Calculate age-specific survivorship from matrix population model

This function uses age-from-stage decomposition methods to calculate age-specific survivorship from a stage-classified matrix population model. A detailed description of these methods can be found in section 5.3.1 of Caswell (2001).

ageSpecificSurv(matU, startLife, N)

Arguments

matU
The survival component of a matrix population model (i.e. a square projection matrix reflecting survival-related transitions; e.g. progression, stasis, and retrogression)

startLife
The index of the first stage at which the author considers the beginning of life.

N
Maximum age to which age-specific survivorship will be calculated.

Value

A vector containing age-specific survivorship, from ages 0 through N.

References

Caswell, H. (2001) Matrix Population Models: Construction, Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

Author

Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>

Examples

matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

ageSpecificSurv(matU, startLife = 1, N = 20)

#>  [1] 1.000000000 0.600000000 0.310000000 0.216000000 0.159100000 0.117360000
#>  [7] 0.086311000 0.063357600 0.046465510 0.034062456 0.024965223 0.018295982
#> [13] 0.013407818 0.009825446 0.007200170 0.005276324 0.003866513 0.002833395
#> [19] 0.002076321 0.001521535 0.001114986