Calculate Demetrius' entropy from trajectories of age-specific survivorship and fecundity

This function calculates Demetrius' entropy from vectors of age-specific survivorship (lx) and fecundity (mx).

entropy_d(lx, mx)

Arguments

lx
Age-specific survivorship trajectory (a vector of monotonically-declining values in the interval [0,1]).

mx
Age-specific fecundity trajectory (a vector of non-negative values).

Value

Demetrius' entropy.

Details

#' @section Warning: Note that this function may produce unexpected results if used on partial survivorship and fecundity trajectories. In addition, it is sensitive to the length of the these vectors. We direct users to the functions `shape_surv` and `shape_rep` which are relatively robust to these issues.

References

Demetrius, L. 1978 Adaptive value, entropy and survivorship curves. Nature, 275, 213-214. <doi:10.1038/275213a0>

See also

Author

Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>

Patrick Barks <patrick.barks@gmail.com>

Examples

data(mpm1)

# derive trajectories of lx and mx, starting from stage 2
lx <- mpm_to_lx(mpm1$matU, start = 2)
mx <- mpm_to_mx(mpm1$matU, mpm1$matF, start = 2)

# calculate Demetrius' entropy
entropy_d(lx, mx)

#> [1] -1.358418