Calculate Keyfitz entropy for an age-based matrix population model

Computes Keyfitz entropy from the U submatrix of an age-based matrix population model.

entropy_k_age(Umat)

Arguments

Umat

A square numeric matrix representing the U submatrix of a matrix population model. The dimension corresponds to the number of age classes

Value

Returns a single numeric value representing the Keyfitz entropy for the given matrix. This value quantifies the dispersion of age at death.

References

Keyfitz, N. 1977. Applied Mathematical Demography. New York: Wiley.

de Vries, C., Bernard, C., & Salguero-Gómez, R. 2023. Discretising Keyfitz' entropy for studies of actuarial senescence and comparative demography. Methods in Ecology and Evolution, 14, 1312–1319. <https://doi.org/10.1111/2041-210X.14083>

See also

Author

Lotte de Vries <c.devries@uva.nl>

Owen Jones <jones@biology.sdu.dk>

Examples

data(leslie_mpm1)

entropy_k_age(leslie_mpm1$matU)
#> [1] 0.626804