Aggregate a Leslie matrix

Takes a Leslie matrix and aggregates it to a desired dimension m using least squares weights equal to the stable age distribution. The output includes the aggregated matrix, the weight matrix, the original (or expanded) Leslie matrix raised to the k power, the partitioning matrix, the size of the original (or expanded) Leslie matrix, the size of the aggregated matrix, the quotient of the original size divided by the aggregated size, and the effectiveness of aggregation.

leslie_collapse(A, m)

Arguments

A

a Leslie matrix

m

the dimensionality of the desired aggregated matrix

Value

a list including the following elements:

B

The aggregated matrix

W

The weight matrix

Ak

The original (or expanded) Leslie matrix raised to the k power

S

The partitioning matrix

n

The size of the original (or expanded) Leslie matrix

m

The size of the aggregated matrix

k

The quotient of the original size divided by the aggregated size

EFF

The effectiveness of aggregation

References

Hinrichsen, R. A. (2023). Aggregation of Leslie matrix models with application to ten diverse animal species. Population Ecology, 1–21. https://doi.org/10.1002/1438-390X.12149

See also

Author

Richard A. Hinrichsen <rich@hinrichsenenvironmental.com>

Examples

data(leslie_mpm1)
A <- leslie_mpm1$matU + leslie_mpm1$matF
leslie_collapse(A, 4)
#> $B
#>           [,1]     [,2]      [,3]       [,4]
#> [1,] 1.3590360 8.062029 6.9520840 5.59842330
#> [2,] 0.6989141 0.000000 0.0000000 0.00000000
#> [3,] 0.0000000 0.458560 0.0000000 0.00000000
#> [4,] 0.0000000 0.000000 0.1887098 0.03029564
#> 
#> $W
#>           [,1]      [,2]      [,3]       [,4]       [,5]       [,6]        [,7]
#> [1,] 0.8781639 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.000000000
#> [2,] 0.0000000 0.4276144 0.0000000 0.00000000 0.00000000 0.00000000 0.000000000
#> [3,] 0.0000000 0.0000000 0.1957251 0.00000000 0.00000000 0.00000000 0.000000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.08181476 0.00000000 0.00000000 0.000000000
#> [5,] 0.0000000 0.0000000 0.0000000 0.00000000 0.02982285 0.00000000 0.000000000
#> [6,] 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.00888094 0.000000000
#> [7,] 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.001949207
#> [8,] 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000 0.000000000
#>              [,8]
#> [1,] 0.0000000000
#> [2,] 0.0000000000
#> [3,] 0.0000000000
#> [4,] 0.0000000000
#> [5,] 0.0000000000
#> [6,] 0.0000000000
#> [7,] 0.0000000000
#> [8,] 0.0002926078
#> 
#> $Ak
#>         0       1        2       3       4        5        6        7
#> 0 0.00000 4.15000 3.790000 3.30500 2.70000 1.990000 1.265000 0.640000
#> 1 0.00000 0.00000 4.415000 4.41500 4.41500 4.415000 4.415000 4.415000
#> 2 0.73289 0.00000 0.000000 0.00000 0.00000 0.000000 0.000000 0.000000
#> 3 0.00000 0.62914 0.000000 0.00000 0.00000 0.000000 0.000000 0.000000
#> 4 0.00000 0.00000 0.501038 0.00000 0.00000 0.000000 0.000000 0.000000
#> 5 0.00000 0.00000 0.000000 0.35694 0.00000 0.000000 0.000000 0.000000
#> 6 0.00000 0.00000 0.000000 0.00000 0.21492 0.000000 0.000000 0.000000
#> 7 0.00000 0.00000 0.000000 0.00000 0.00000 0.100694 0.032384 0.016384
#> 
#> $S
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,]    1    1    0    0    0    0    0    0
#> [2,]    0    0    1    1    0    0    0    0
#> [3,]    0    0    0    0    1    1    0    0
#> [4,]    0    0    0    0    0    0    1    1
#> 
#> $n
#> [1] 8
#> 
#> $m
#> [1] 4
#> 
#> $k
#> [1] 2
#> 
#> $EFF
#> [1] 0.8227475
#>