Calculate the timing of lifetime reproductive events

Applies Markov chain approaches to decompose various moments along the age-specific trajectory of reproduction of individuals in a matrix population model, including the probability of achieving maturity, age at first reproduction, mean life expectancy conditional on maturity, and life expectancy for mature individuals.

lifeTimeRepEvents(matU, matR, startLife = 1)

Arguments

matU
The survival component of a matrix population model (i.e. a square projection matrix reflecting survival-related transitions; e.g. progression, stasis, and retrogression)

matR
The reproductive component of a matrix population model (i.e. a square projection matrix reflecting transitions due to reproduction; either sexual, clonal, or both)

startLife
The index of the first stage at which the author considers the beginning of life. Defaults to 1.

Value

A list with four elements:

p

probability of achieving reproductive maturity

La

mean age at maturity (in the same units as the matrix population model sampling periodicity)

meanLifeExpectancy

mean life expectancy conditional on entering the life cycle at the value of startLife

remainingMatureLifeExpectancy

Life expectancy from mean maturity, calculated as meanLifeExpectancy - La. This value can be negative because both mean life expectancy and mean age at maturity are means of their respective distributions, and their distributions can indeed overlap

References

Caswell, H. (2001) Matrix Population Models: Construction, Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

Author

Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <hcaswell@whoi.edu>

Owen R. Jones <jones@biology.sdu.dk>

Examples

matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

lifeTimeRepEvents(matU, matF, startLife = 1)

#> $p
#> [1] 0.2083333
#> 
#> $La
#> [1] 3.361111
#> 
#> $meanLifeExpectancy
#> [1] 4.418605
#> 
#> $remainingMatureLifeExpectancy
#> [1] -0.6350129
#>