Generate a life table from a matrix population model

This function uses age-from-stage decomposition methods to generate a life table from a matrix population model. A detailed description of these methods can be found in section 5.3 of Caswell (2001).

makeLifeTable(matU, matF = NULL, matC = NULL, startLife = 1,
  nSteps = 1000)

Arguments

matU
The survival component of a matrix population model (i.e. a square projection matrix reflecting survival-related transitions; e.g. progression, stasis, and retrogression)

matF
(Optional) The sexual component of a matrix population model (i.e. a square projection matrix reflecting transitions due to sexual reproduction)

matC
(Optional) The clonal component of a matrix population model (i.e. a square projection matrix reflecting transitions due to clonal reproduction).

startLife
The index of the first stage at which the author considers the beginning of life. Defaults to 1.

nSteps
Number of time steps for which the life table will be calculated. Time steps are in the same units as the matrix population model (see MatrixPeriodicity column in metadata of COM(P)ADRE). Defaults to 1000.

Value

A data.frame containing 7-13 columns and nSteps rows. Columns include:

x

age

lx

survivorship to start of age class x

dx

proportion of original cohort dying in interval [x, x+1)

qx

force of mortality at age x

Lx

survivorship to middle of age class x

Tx

proportion of original cohort alive at age x and beyond

ex

remaining life expectancy at age x

If matF is provided, also includes:
mx

per-capita rate of sexual reproduction at age x

lxmx

expected number of sexual offspring per original cohort member produced at age x

If matC is provided, also includes:
cx

per-capita rate of clonal reproduction at age x

lxcx

expected number of clonal offspring per original cohort member produced at age x

If both matF and matC are provided, also includes:
mxcx

per-capita rate of total reproduction (sexual + clonal) at age x

lxmxcx

expected number of total offspring (sexual + clonal) per original cohort member produced at age x

References

Caswell, H. (2001) Matrix Population Models: Construction, Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

Caswell, H. (2006) Applications of Markov chains in demography. pp. 319-334 in A.N. Langville and W.J. Stewart (editors) MAM2006: Markov Anniversary Meeting. Boson Books, Raleigh, North Caroline, USA

Jones, O.R. et al. (2014) Diversity of ageing across the tree of life. Nature, 505(7482), 169–173

Author

Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>

Examples

matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))
              
matC <- rbind(c(  0,   0, 0.4, 0.5),
              c(  0,   0, 0.3, 0.1),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

makeLifeTable(matU, startLife = 1, nSteps = 100)

#>      x           lx           dx        qx           Lx           Tx        ex
#> 1    0 1.000000e+00 4.000000e-01 0.4000000 8.000000e-01 2.226098e+00 2.2260982
#> 2    1 6.000000e-01 2.900000e-01 0.4833333 4.550000e-01 1.426098e+00 2.3768303
#> 3    2 3.100000e-01 9.400000e-02 0.3032258 2.630000e-01 9.710982e-01 3.1325748
#> 4    3 2.160000e-01 5.690000e-02 0.2634259 1.875500e-01 7.080982e-01 3.2782324
#> 5    4 1.591000e-01 4.174000e-02 0.2623507 1.382300e-01 5.205482e-01 3.2718302
#> 6    5 1.173600e-01 3.104900e-02 0.2645620 1.018355e-01 3.823182e-01 3.2576533
#> 7    6 8.631100e-02 2.295340e-02 0.2659383 7.483430e-02 2.804827e-01 3.2496749
#> 8    7 6.335760e-02 1.689209e-02 0.2666151 5.491156e-02 2.056484e-01 3.2458362
#> 9    8 4.646551e-02 1.240305e-02 0.2669303 4.026398e-02 1.507368e-01 3.2440586
#> 10   9 3.406246e-02 9.097233e-03 0.2670751 2.951384e-02 1.104729e-01 3.2432439
#> 11  10 2.496522e-02 6.669241e-03 0.2671413 2.163060e-02 8.095901e-02 3.2428716
#> 12  11 1.829598e-02 4.888165e-03 0.2671715 1.585190e-02 5.932841e-02 3.2427016
#> 13  12 1.340782e-02 3.582371e-03 0.2671853 1.161663e-02 4.347651e-02 3.2426240
#> 14  13 9.825446e-03 2.625276e-03 0.2671916 8.512808e-03 3.185988e-02 3.2425886
#> 15  14 7.200170e-03 1.923845e-03 0.2671945 6.238247e-03 2.334707e-02 3.2425724
#> 16  15 5.276324e-03 1.409812e-03 0.2671958 4.571419e-03 1.710882e-02 3.2425650
#> 17  16 3.866513e-03 1.033118e-03 0.2671964 3.349954e-03 1.253741e-02 3.2425616
#> 18  17 2.833395e-03 7.570735e-04 0.2671966 2.454858e-03 9.187452e-03 3.2425601
#> 19  18 2.076321e-03 5.547863e-04 0.2671968 1.798928e-03 6.732594e-03 3.2425594
#> 20  19 1.521535e-03 4.065493e-04 0.2671968 1.318260e-03 4.933666e-03 3.2425591
#> 21  20 1.114986e-03 2.979206e-04 0.2671968 9.660252e-04 3.615406e-03 3.2425589
#> 22  21 8.170649e-04 2.183172e-04 0.2671969 7.079063e-04 2.649381e-03 3.2425589
#> 23  22 5.987477e-04 1.599835e-04 0.2671969 5.187560e-04 1.941475e-03 3.2425588
#> 24  23 4.387642e-04 1.172364e-04 0.2671969 3.801460e-04 1.422719e-03 3.2425588
#> 25  24 3.215278e-04 8.591122e-05 0.2671969 2.785722e-04 1.042573e-03 3.2425588
#> 26  25 2.356166e-04 6.295601e-05 0.2671969 2.041386e-04 7.640006e-04 3.2425588
#> 27  26 1.726606e-04 4.613436e-05 0.2671969 1.495934e-04 5.598620e-04 3.2425588
#> 28  27 1.265262e-04 3.380740e-05 0.2671969 1.096225e-04 4.102686e-04 3.2425588
#> 29  28 9.271879e-05 2.477417e-05 0.2671969 8.033171e-05 3.006461e-04 3.2425588
#> 30  29 6.794462e-05 1.815459e-05 0.2671969 5.886733e-05 2.203144e-04 3.2425588
#> 31  30 4.979003e-05 1.330374e-05 0.2671969 4.313816e-05 1.614471e-04 3.2425588
#> 32  31 3.648629e-05 9.749023e-06 0.2671969 3.161178e-05 1.183089e-04 3.2425588
#> 33  32 2.673727e-05 7.144114e-06 0.2671969 2.316521e-05 8.669717e-05 3.2425588
#> 34  33 1.959315e-05 5.235229e-06 0.2671969 1.697554e-05 6.353195e-05 3.2425588
#> 35  34 1.435792e-05 3.836393e-06 0.2671969 1.243973e-05 4.655641e-05 3.2425588
#> 36  35 1.052153e-05 2.811320e-06 0.2671969 9.115872e-06 3.411669e-05 3.2425588
#> 37  36 7.710212e-06 2.060144e-06 0.2671969 6.680139e-06 2.500081e-05 3.2425588
#> 38  37 5.650067e-06 1.509680e-06 0.2671969 4.895227e-06 1.832067e-05 3.2425588
#> 39  38 4.140387e-06 1.106298e-06 0.2671969 3.587238e-06 1.342545e-05 3.2425588
#> 40  39 3.034088e-06 8.106989e-07 0.2671969 2.628739e-06 9.838210e-06 3.2425588
#> 41  40 2.223390e-06 5.940827e-07 0.2671969 1.926348e-06 7.209471e-06 3.2425588
#> 42  41 1.629307e-06 4.353457e-07 0.2671969 1.411634e-06 5.283123e-06 3.2425587
#> 43  42 1.193961e-06 3.190227e-07 0.2671969 1.034450e-06 3.871489e-06 3.2425587
#> 44  43 8.749385e-07 2.337808e-07 0.2671969 7.580480e-07 2.837039e-06 3.2425587
#> 45  44 6.411576e-07 1.713153e-07 0.2671969 5.555000e-07 2.078991e-06 3.2425587
#> 46  45 4.698423e-07 1.255404e-07 0.2671969 4.070721e-07 1.523491e-06 3.2425586
#> 47  46 3.443019e-07 9.199640e-08 0.2671969 2.983037e-07 1.116419e-06 3.2425586
#> 48  47 2.523055e-07 6.741525e-08 0.2671969 2.185979e-07 8.181154e-07 3.2425585
#> 49  48 1.848903e-07 4.940210e-08 0.2671969 1.601892e-07 5.995175e-07 3.2425584
#> 50  49 1.354882e-07 3.620202e-08 0.2671969 1.173872e-07 4.393283e-07 3.2425582
#> 51  50 9.928616e-08 2.652895e-08 0.2671969 8.602168e-08 3.219411e-07 3.2425580
#> 52  51 7.275721e-08 1.944050e-08 0.2671969 6.303696e-08 2.359194e-07 3.2425577
#> 53  52 5.331671e-08 1.424606e-08 0.2671969 4.619368e-08 1.728825e-07 3.2425573
#> 54  53 3.907065e-08 1.043956e-08 0.2671969 3.385087e-08 1.266888e-07 3.2425568
#> 55  54 2.863110e-08 7.650139e-09 0.2671969 2.480603e-08 9.283793e-08 3.2425561
#> 56  55 2.098096e-08 5.606046e-09 0.2671969 1.817793e-08 6.803191e-08 3.2425551
#> 57  56 1.537491e-08 4.108128e-09 0.2671969 1.332085e-08 4.985397e-08 3.2425537
#> 58  57 1.126678e-08 3.010449e-09 0.2671969 9.761558e-09 3.653313e-08 3.2425519
#> 59  58 8.256334e-09 2.206066e-09 0.2671969 7.153300e-09 2.677157e-08 3.2425493
#> 60  59 6.050267e-09 1.616612e-09 0.2671969 5.241961e-09 1.961827e-08 3.2425459
#> 61  60 4.433655e-09 1.184659e-09 0.2671969 3.841325e-09 1.437631e-08 3.2425412
#> 62  61 3.248996e-09 8.681216e-10 0.2671969 2.814935e-09 1.053498e-08 3.2425348
#> 63  62 2.380874e-09 6.361622e-10 0.2671969 2.062793e-09 7.720047e-09 3.2425260
#> 64  63 1.744712e-09 4.661816e-10 0.2671969 1.511621e-09 5.657254e-09 3.2425141
#> 65  64 1.278531e-09 3.416194e-10 0.2671969 1.107721e-09 4.145633e-09 3.2424978
#> 66  65 9.369112e-10 2.503397e-10 0.2671969 8.117413e-10 3.037912e-09 3.2424755
#> 67  66 6.865715e-10 1.834497e-10 0.2671969 5.948466e-10 2.226170e-09 3.2424452
#> 68  67 5.031217e-10 1.344325e-10 0.2671969 4.359054e-10 1.631324e-09 3.2424037
#> 69  68 3.686892e-10 9.851259e-11 0.2671969 3.194329e-10 1.195418e-09 3.2423472
#> 70  69 2.701766e-10 7.219034e-11 0.2671969 2.340814e-10 8.759854e-10 3.2422701
#> 71  70 1.979862e-10 5.290130e-11 0.2671969 1.715356e-10 6.419040e-10 3.2421648
#> 72  71 1.450849e-10 3.876624e-11 0.2671969 1.257018e-10 4.703684e-10 3.2420211
#> 73  72 1.063187e-10 2.840802e-11 0.2671969 9.211468e-11 3.446666e-10 3.2418251
#> 74  73 7.791067e-11 2.081749e-11 0.2671969 6.750193e-11 2.525519e-10 3.2415575
#> 75  74 5.709318e-11 1.525512e-11 0.2671969 4.946562e-11 1.850500e-10 3.2411924
#> 76  75 4.183806e-11 1.117900e-11 0.2671969 3.624856e-11 1.355844e-10 3.2406942
#> 77  76 3.065906e-11 8.192006e-12 0.2671969 2.656306e-11 9.933581e-11 3.2400143
#> 78  77 2.246706e-11 6.003128e-12 0.2671969 1.946549e-11 7.277275e-11 3.2390866
#> 79  78 1.646393e-11 4.399111e-12 0.2671969 1.426438e-11 5.330725e-11 3.2378205
#> 80  79 1.206482e-11 3.223682e-12 0.2671969 1.045298e-11 3.904288e-11 3.2360928
#> 81  80 8.841138e-12 2.362324e-12 0.2671969 7.659976e-12 2.858990e-11 3.2337352
#> 82  81 6.478813e-12 1.731119e-12 0.2671969 5.613254e-12 2.092992e-11 3.2305179
#> 83  82 4.747695e-12 1.268569e-12 0.2671969 4.113410e-12 1.531667e-11 3.2261275
#> 84  83 3.479126e-12 9.296115e-13 0.2671969 3.014320e-12 1.120326e-11 3.2201363
#> 85  84 2.549514e-12 6.812222e-13 0.2671969 2.208903e-12 8.188939e-12 3.2119605
#> 86  85 1.868292e-12 4.992018e-13 0.2671969 1.618691e-12 5.980036e-12 3.2008036
#> 87  86 1.369090e-12 3.658166e-13 0.2671969 1.186182e-12 4.361344e-12 3.1855787
#> 88  87 1.003274e-12 2.680716e-13 0.2671969 8.692378e-13 3.175163e-12 3.1648025
#> 89  88 7.352020e-13 1.964437e-13 0.2671969 6.369802e-13 2.305925e-12 3.1364507
#> 90  89 5.387583e-13 1.439545e-13 0.2671969 4.667811e-13 1.668945e-12 3.0977613
#> 91  90 3.948038e-13 1.054903e-13 0.2671969 3.420586e-13 1.202164e-12 3.0449648
#> 92  91 2.893134e-13 7.730365e-14 0.2671969 2.506616e-13 8.601050e-13 2.9729175
#> 93  92 2.120098e-13 5.664836e-14 0.2671969 1.836856e-13 6.094434e-13 2.8746001
#> 94  93 1.553614e-13 4.151209e-14 0.2671969 1.346054e-13 4.257578e-13 2.7404339
#> 95  94 1.138494e-13 3.042019e-14 0.2671969 9.863926e-14 2.911524e-13 2.5573476
#> 96  95 8.342916e-14 2.229201e-14 0.2671969 7.228316e-14 1.925131e-13 2.3075039
#> 97  96 6.113715e-14 1.633566e-14 0.2671969 5.296932e-14 1.202300e-13 1.9665614
#> 98  97 4.480150e-14 1.197082e-14 0.2671969 3.881609e-14 6.726064e-14 1.5013033
#> 99  98 3.283068e-14 8.772254e-15 0.2671969 2.844455e-14 2.844455e-14 0.8664016
#> 100 99 2.405842e-14           NA        NA           NA           NA        NA

makeLifeTable(matU, matF, startLife = 1, nSteps = 100)

#>      x           lx           dx        qx           Lx           Tx        ex
#> 1    0 1.000000e+00 4.000000e-01 0.4000000 8.000000e-01 2.226098e+00 2.2260982
#> 2    1 6.000000e-01 2.900000e-01 0.4833333 4.550000e-01 1.426098e+00 2.3768303
#> 3    2 3.100000e-01 9.400000e-02 0.3032258 2.630000e-01 9.710982e-01 3.1325748
#> 4    3 2.160000e-01 5.690000e-02 0.2634259 1.875500e-01 7.080982e-01 3.2782324
#> 5    4 1.591000e-01 4.174000e-02 0.2623507 1.382300e-01 5.205482e-01 3.2718302
#> 6    5 1.173600e-01 3.104900e-02 0.2645620 1.018355e-01 3.823182e-01 3.2576533
#> 7    6 8.631100e-02 2.295340e-02 0.2659383 7.483430e-02 2.804827e-01 3.2496749
#> 8    7 6.335760e-02 1.689209e-02 0.2666151 5.491156e-02 2.056484e-01 3.2458362
#> 9    8 4.646551e-02 1.240305e-02 0.2669303 4.026398e-02 1.507368e-01 3.2440586
#> 10   9 3.406246e-02 9.097233e-03 0.2670751 2.951384e-02 1.104729e-01 3.2432439
#> 11  10 2.496522e-02 6.669241e-03 0.2671413 2.163060e-02 8.095901e-02 3.2428716
#> 12  11 1.829598e-02 4.888165e-03 0.2671715 1.585190e-02 5.932841e-02 3.2427016
#> 13  12 1.340782e-02 3.582371e-03 0.2671853 1.161663e-02 4.347651e-02 3.2426240
#> 14  13 9.825446e-03 2.625276e-03 0.2671916 8.512808e-03 3.185988e-02 3.2425886
#> 15  14 7.200170e-03 1.923845e-03 0.2671945 6.238247e-03 2.334707e-02 3.2425724
#> 16  15 5.276324e-03 1.409812e-03 0.2671958 4.571419e-03 1.710882e-02 3.2425650
#> 17  16 3.866513e-03 1.033118e-03 0.2671964 3.349954e-03 1.253741e-02 3.2425616
#> 18  17 2.833395e-03 7.570735e-04 0.2671966 2.454858e-03 9.187452e-03 3.2425601
#> 19  18 2.076321e-03 5.547863e-04 0.2671968 1.798928e-03 6.732594e-03 3.2425594
#> 20  19 1.521535e-03 4.065493e-04 0.2671968 1.318260e-03 4.933666e-03 3.2425591
#> 21  20 1.114986e-03 2.979206e-04 0.2671968 9.660252e-04 3.615406e-03 3.2425589
#> 22  21 8.170649e-04 2.183172e-04 0.2671969 7.079063e-04 2.649381e-03 3.2425589
#> 23  22 5.987477e-04 1.599835e-04 0.2671969 5.187560e-04 1.941475e-03 3.2425588
#> 24  23 4.387642e-04 1.172364e-04 0.2671969 3.801460e-04 1.422719e-03 3.2425588
#> 25  24 3.215278e-04 8.591122e-05 0.2671969 2.785722e-04 1.042573e-03 3.2425588
#> 26  25 2.356166e-04 6.295601e-05 0.2671969 2.041386e-04 7.640006e-04 3.2425588
#> 27  26 1.726606e-04 4.613436e-05 0.2671969 1.495934e-04 5.598620e-04 3.2425588
#> 28  27 1.265262e-04 3.380740e-05 0.2671969 1.096225e-04 4.102686e-04 3.2425588
#> 29  28 9.271879e-05 2.477417e-05 0.2671969 8.033171e-05 3.006461e-04 3.2425588
#> 30  29 6.794462e-05 1.815459e-05 0.2671969 5.886733e-05 2.203144e-04 3.2425588
#> 31  30 4.979003e-05 1.330374e-05 0.2671969 4.313816e-05 1.614471e-04 3.2425588
#> 32  31 3.648629e-05 9.749023e-06 0.2671969 3.161178e-05 1.183089e-04 3.2425588
#> 33  32 2.673727e-05 7.144114e-06 0.2671969 2.316521e-05 8.669717e-05 3.2425588
#> 34  33 1.959315e-05 5.235229e-06 0.2671969 1.697554e-05 6.353195e-05 3.2425588
#> 35  34 1.435792e-05 3.836393e-06 0.2671969 1.243973e-05 4.655641e-05 3.2425588
#> 36  35 1.052153e-05 2.811320e-06 0.2671969 9.115872e-06 3.411669e-05 3.2425588
#> 37  36 7.710212e-06 2.060144e-06 0.2671969 6.680139e-06 2.500081e-05 3.2425588
#> 38  37 5.650067e-06 1.509680e-06 0.2671969 4.895227e-06 1.832067e-05 3.2425588
#> 39  38 4.140387e-06 1.106298e-06 0.2671969 3.587238e-06 1.342545e-05 3.2425588
#> 40  39 3.034088e-06 8.106989e-07 0.2671969 2.628739e-06 9.838210e-06 3.2425588
#> 41  40 2.223390e-06 5.940827e-07 0.2671969 1.926348e-06 7.209471e-06 3.2425588
#> 42  41 1.629307e-06 4.353457e-07 0.2671969 1.411634e-06 5.283123e-06 3.2425587
#> 43  42 1.193961e-06 3.190227e-07 0.2671969 1.034450e-06 3.871489e-06 3.2425587
#> 44  43 8.749385e-07 2.337808e-07 0.2671969 7.580480e-07 2.837039e-06 3.2425587
#> 45  44 6.411576e-07 1.713153e-07 0.2671969 5.555000e-07 2.078991e-06 3.2425587
#> 46  45 4.698423e-07 1.255404e-07 0.2671969 4.070721e-07 1.523491e-06 3.2425586
#> 47  46 3.443019e-07 9.199640e-08 0.2671969 2.983037e-07 1.116419e-06 3.2425586
#> 48  47 2.523055e-07 6.741525e-08 0.2671969 2.185979e-07 8.181154e-07 3.2425585
#> 49  48 1.848903e-07 4.940210e-08 0.2671969 1.601892e-07 5.995175e-07 3.2425584
#> 50  49 1.354882e-07 3.620202e-08 0.2671969 1.173872e-07 4.393283e-07 3.2425582
#> 51  50 9.928616e-08 2.652895e-08 0.2671969 8.602168e-08 3.219411e-07 3.2425580
#> 52  51 7.275721e-08 1.944050e-08 0.2671969 6.303696e-08 2.359194e-07 3.2425577
#> 53  52 5.331671e-08 1.424606e-08 0.2671969 4.619368e-08 1.728825e-07 3.2425573
#> 54  53 3.907065e-08 1.043956e-08 0.2671969 3.385087e-08 1.266888e-07 3.2425568
#> 55  54 2.863110e-08 7.650139e-09 0.2671969 2.480603e-08 9.283793e-08 3.2425561
#> 56  55 2.098096e-08 5.606046e-09 0.2671969 1.817793e-08 6.803191e-08 3.2425551
#> 57  56 1.537491e-08 4.108128e-09 0.2671969 1.332085e-08 4.985397e-08 3.2425537
#> 58  57 1.126678e-08 3.010449e-09 0.2671969 9.761558e-09 3.653313e-08 3.2425519
#> 59  58 8.256334e-09 2.206066e-09 0.2671969 7.153300e-09 2.677157e-08 3.2425493
#> 60  59 6.050267e-09 1.616612e-09 0.2671969 5.241961e-09 1.961827e-08 3.2425459
#> 61  60 4.433655e-09 1.184659e-09 0.2671969 3.841325e-09 1.437631e-08 3.2425412
#> 62  61 3.248996e-09 8.681216e-10 0.2671969 2.814935e-09 1.053498e-08 3.2425348
#> 63  62 2.380874e-09 6.361622e-10 0.2671969 2.062793e-09 7.720047e-09 3.2425260
#> 64  63 1.744712e-09 4.661816e-10 0.2671969 1.511621e-09 5.657254e-09 3.2425141
#> 65  64 1.278531e-09 3.416194e-10 0.2671969 1.107721e-09 4.145633e-09 3.2424978
#> 66  65 9.369112e-10 2.503397e-10 0.2671969 8.117413e-10 3.037912e-09 3.2424755
#> 67  66 6.865715e-10 1.834497e-10 0.2671969 5.948466e-10 2.226170e-09 3.2424452
#> 68  67 5.031217e-10 1.344325e-10 0.2671969 4.359054e-10 1.631324e-09 3.2424037
#> 69  68 3.686892e-10 9.851259e-11 0.2671969 3.194329e-10 1.195418e-09 3.2423472
#> 70  69 2.701766e-10 7.219034e-11 0.2671969 2.340814e-10 8.759854e-10 3.2422701
#> 71  70 1.979862e-10 5.290130e-11 0.2671969 1.715356e-10 6.419040e-10 3.2421648
#> 72  71 1.450849e-10 3.876624e-11 0.2671969 1.257018e-10 4.703684e-10 3.2420211
#> 73  72 1.063187e-10 2.840802e-11 0.2671969 9.211468e-11 3.446666e-10 3.2418251
#> 74  73 7.791067e-11 2.081749e-11 0.2671969 6.750193e-11 2.525519e-10 3.2415575
#> 75  74 5.709318e-11 1.525512e-11 0.2671969 4.946562e-11 1.850500e-10 3.2411924
#> 76  75 4.183806e-11 1.117900e-11 0.2671969 3.624856e-11 1.355844e-10 3.2406942
#> 77  76 3.065906e-11 8.192006e-12 0.2671969 2.656306e-11 9.933581e-11 3.2400143
#> 78  77 2.246706e-11 6.003128e-12 0.2671969 1.946549e-11 7.277275e-11 3.2390866
#> 79  78 1.646393e-11 4.399111e-12 0.2671969 1.426438e-11 5.330725e-11 3.2378205
#> 80  79 1.206482e-11 3.223682e-12 0.2671969 1.045298e-11 3.904288e-11 3.2360928
#> 81  80 8.841138e-12 2.362324e-12 0.2671969 7.659976e-12 2.858990e-11 3.2337352
#> 82  81 6.478813e-12 1.731119e-12 0.2671969 5.613254e-12 2.092992e-11 3.2305179
#> 83  82 4.747695e-12 1.268569e-12 0.2671969 4.113410e-12 1.531667e-11 3.2261275
#> 84  83 3.479126e-12 9.296115e-13 0.2671969 3.014320e-12 1.120326e-11 3.2201363
#> 85  84 2.549514e-12 6.812222e-13 0.2671969 2.208903e-12 8.188939e-12 3.2119605
#> 86  85 1.868292e-12 4.992018e-13 0.2671969 1.618691e-12 5.980036e-12 3.2008036
#> 87  86 1.369090e-12 3.658166e-13 0.2671969 1.186182e-12 4.361344e-12 3.1855787
#> 88  87 1.003274e-12 2.680716e-13 0.2671969 8.692378e-13 3.175163e-12 3.1648025
#> 89  88 7.352020e-13 1.964437e-13 0.2671969 6.369802e-13 2.305925e-12 3.1364507
#> 90  89 5.387583e-13 1.439545e-13 0.2671969 4.667811e-13 1.668945e-12 3.0977613
#> 91  90 3.948038e-13 1.054903e-13 0.2671969 3.420586e-13 1.202164e-12 3.0449648
#> 92  91 2.893134e-13 7.730365e-14 0.2671969 2.506616e-13 8.601050e-13 2.9729175
#> 93  92 2.120098e-13 5.664836e-14 0.2671969 1.836856e-13 6.094434e-13 2.8746001
#> 94  93 1.553614e-13 4.151209e-14 0.2671969 1.346054e-13 4.257578e-13 2.7404339
#> 95  94 1.138494e-13 3.042019e-14 0.2671969 9.863926e-14 2.911524e-13 2.5573476
#> 96  95 8.342916e-14 2.229201e-14 0.2671969 7.228316e-14 1.925131e-13 2.3075039
#> 97  96 6.113715e-14 1.633566e-14 0.2671969 5.296932e-14 1.202300e-13 1.9665614
#> 98  97 4.480150e-14 1.197082e-14 0.2671969 3.881609e-14 6.726064e-14 1.5013033
#> 99  98 3.283068e-14 8.772254e-15 0.2671969 2.844455e-14 2.844455e-14 0.8664016
#> 100 99 2.405842e-14           NA        NA           NA           NA        NA
#>            mx         lxmx
#> 1   0.0000000 0.000000e+00
#> 2   0.0000000 0.000000e+00
#> 3   0.9193548 2.850000e-01
#> 4   1.4861111 3.210000e-01
#> 5   1.7225016 2.740500e-01
#> 6   1.8228528 2.139300e-01
#> 7   1.8672533 1.611645e-01
#> 8   1.8872827 1.195737e-01
#> 9   1.8963856 8.811653e-02
#> 10  1.9005343 6.473687e-02
#> 11  1.9024270 4.749451e-02
#> 12  1.9032909 3.482258e-02
#> 13  1.9036853 2.552426e-02
#> 14  1.9038653 1.870633e-02
#> 15  1.9039475 1.370875e-02
#> 16  1.9039850 1.004604e-02
#> 17  1.9040022 7.361849e-03
#> 18  1.9040100 5.394812e-03
#> 19  1.9040136 3.953343e-03
#> 20  1.9040152 2.897025e-03
#> 21  1.9040159 2.122950e-03
#> 22  1.9040163 1.555705e-03
#> 23  1.9040164 1.140026e-03
#> 24  1.9040165 8.354143e-04
#> 25  1.9040165 6.121942e-04
#> 26  1.9040165 4.486178e-04
#> 27  1.9040166 3.287486e-04
#> 28  1.9040166 2.409080e-04
#> 29  1.9040166 1.765381e-04
#> 30  1.9040166 1.293677e-04
#> 31  1.9040166 9.480105e-05
#> 32  1.9040166 6.947050e-05
#> 33  1.9040166 5.090820e-05
#> 34  1.9040166 3.730569e-05
#> 35  1.9040166 2.733773e-05
#> 36  1.9040166 2.003317e-05
#> 37  1.9040166 1.468037e-05
#> 38  1.9040166 1.075782e-05
#> 39  1.9040166 7.883365e-06
#> 40  1.9040166 5.776955e-06
#> 41  1.9040166 4.233370e-06
#> 42  1.9040166 3.102227e-06
#> 43  1.9040166 2.273322e-06
#> 44  1.9040166 1.665897e-06
#> 45  1.9040166 1.220775e-06
#> 46  1.9040166 8.945876e-07
#> 47  1.9040166 6.555566e-07
#> 48  1.9040166 4.803939e-07
#> 49  1.9040166 3.520342e-07
#> 50  1.9040166 2.579717e-07
#> 51  1.9040166 1.890425e-07
#> 52  1.9040166 1.385309e-07
#> 53  1.9040166 1.015159e-07
#> 54  1.9040166 7.439117e-08
#> 55  1.9040166 5.451408e-08
#> 56  1.9040166 3.994809e-08
#> 57  1.9040166 2.927408e-08
#> 58  1.9040166 2.145214e-08
#> 59  1.9040166 1.572020e-08
#> 60  1.9040166 1.151981e-08
#> 61  1.9040166 8.441752e-09
#> 62  1.9040166 6.186142e-09
#> 63  1.9040166 4.533224e-09
#> 64  1.9040166 3.321961e-09
#> 65  1.9040166 2.434343e-09
#> 66  1.9040166 1.783894e-09
#> 67  1.9040166 1.307243e-09
#> 68  1.9040166 9.579521e-10
#> 69  1.9040166 7.019903e-10
#> 70  1.9040166 5.144207e-10
#> 71  1.9040166 3.769691e-10
#> 72  1.9040166 2.762441e-10
#> 73  1.9040166 2.024326e-10
#> 74  1.9040166 1.483432e-10
#> 75  1.9040166 1.087064e-10
#> 76  1.9040166 7.966037e-11
#> 77  1.9040166 5.837537e-11
#> 78  1.9040166 4.277765e-11
#> 79  1.9040166 3.134760e-11
#> 80  1.9040166 2.297162e-11
#> 81  1.9040166 1.683367e-11
#> 82  1.9040166 1.233577e-11
#> 83  1.9040166 9.039689e-12
#> 84  1.9040166 6.624313e-12
#> 85  1.9040166 4.854317e-12
#> 86  1.9040166 3.557259e-12
#> 87  1.9040166 2.606770e-12
#> 88  1.9040166 1.910249e-12
#> 89  1.9040166 1.399837e-12
#> 90  1.9040166 1.025805e-12
#> 91  1.9040166 7.517129e-13
#> 92  1.9040166 5.508576e-13
#> 93  1.9040166 4.036702e-13
#> 94  1.9040166 2.958108e-13
#> 95  1.9040166 2.167711e-13
#> 96  1.9040166 1.588505e-13
#> 97  1.9040166 1.164061e-13
#> 98  1.9040166 8.530279e-14
#> 99  1.9040166 6.251015e-14
#> 100 1.9040166 4.580763e-14

makeLifeTable(matU, matF, matC, startLife = 1, nSteps = 100)

#>      x           lx           dx        qx           Lx           Tx        ex
#> 1    0 1.000000e+00 4.000000e-01 0.4000000 8.000000e-01 2.226098e+00 2.2260982
#> 2    1 6.000000e-01 2.900000e-01 0.4833333 4.550000e-01 1.426098e+00 2.3768303
#> 3    2 3.100000e-01 9.400000e-02 0.3032258 2.630000e-01 9.710982e-01 3.1325748
#> 4    3 2.160000e-01 5.690000e-02 0.2634259 1.875500e-01 7.080982e-01 3.2782324
#> 5    4 1.591000e-01 4.174000e-02 0.2623507 1.382300e-01 5.205482e-01 3.2718302
#> 6    5 1.173600e-01 3.104900e-02 0.2645620 1.018355e-01 3.823182e-01 3.2576533
#> 7    6 8.631100e-02 2.295340e-02 0.2659383 7.483430e-02 2.804827e-01 3.2496749
#> 8    7 6.335760e-02 1.689209e-02 0.2666151 5.491156e-02 2.056484e-01 3.2458362
#> 9    8 4.646551e-02 1.240305e-02 0.2669303 4.026398e-02 1.507368e-01 3.2440586
#> 10   9 3.406246e-02 9.097233e-03 0.2670751 2.951384e-02 1.104729e-01 3.2432439
#> 11  10 2.496522e-02 6.669241e-03 0.2671413 2.163060e-02 8.095901e-02 3.2428716
#> 12  11 1.829598e-02 4.888165e-03 0.2671715 1.585190e-02 5.932841e-02 3.2427016
#> 13  12 1.340782e-02 3.582371e-03 0.2671853 1.161663e-02 4.347651e-02 3.2426240
#> 14  13 9.825446e-03 2.625276e-03 0.2671916 8.512808e-03 3.185988e-02 3.2425886
#> 15  14 7.200170e-03 1.923845e-03 0.2671945 6.238247e-03 2.334707e-02 3.2425724
#> 16  15 5.276324e-03 1.409812e-03 0.2671958 4.571419e-03 1.710882e-02 3.2425650
#> 17  16 3.866513e-03 1.033118e-03 0.2671964 3.349954e-03 1.253741e-02 3.2425616
#> 18  17 2.833395e-03 7.570735e-04 0.2671966 2.454858e-03 9.187452e-03 3.2425601
#> 19  18 2.076321e-03 5.547863e-04 0.2671968 1.798928e-03 6.732594e-03 3.2425594
#> 20  19 1.521535e-03 4.065493e-04 0.2671968 1.318260e-03 4.933666e-03 3.2425591
#> 21  20 1.114986e-03 2.979206e-04 0.2671968 9.660252e-04 3.615406e-03 3.2425589
#> 22  21 8.170649e-04 2.183172e-04 0.2671969 7.079063e-04 2.649381e-03 3.2425589
#> 23  22 5.987477e-04 1.599835e-04 0.2671969 5.187560e-04 1.941475e-03 3.2425588
#> 24  23 4.387642e-04 1.172364e-04 0.2671969 3.801460e-04 1.422719e-03 3.2425588
#> 25  24 3.215278e-04 8.591122e-05 0.2671969 2.785722e-04 1.042573e-03 3.2425588
#> 26  25 2.356166e-04 6.295601e-05 0.2671969 2.041386e-04 7.640006e-04 3.2425588
#> 27  26 1.726606e-04 4.613436e-05 0.2671969 1.495934e-04 5.598620e-04 3.2425588
#> 28  27 1.265262e-04 3.380740e-05 0.2671969 1.096225e-04 4.102686e-04 3.2425588
#> 29  28 9.271879e-05 2.477417e-05 0.2671969 8.033171e-05 3.006461e-04 3.2425588
#> 30  29 6.794462e-05 1.815459e-05 0.2671969 5.886733e-05 2.203144e-04 3.2425588
#> 31  30 4.979003e-05 1.330374e-05 0.2671969 4.313816e-05 1.614471e-04 3.2425588
#> 32  31 3.648629e-05 9.749023e-06 0.2671969 3.161178e-05 1.183089e-04 3.2425588
#> 33  32 2.673727e-05 7.144114e-06 0.2671969 2.316521e-05 8.669717e-05 3.2425588
#> 34  33 1.959315e-05 5.235229e-06 0.2671969 1.697554e-05 6.353195e-05 3.2425588
#> 35  34 1.435792e-05 3.836393e-06 0.2671969 1.243973e-05 4.655641e-05 3.2425588
#> 36  35 1.052153e-05 2.811320e-06 0.2671969 9.115872e-06 3.411669e-05 3.2425588
#> 37  36 7.710212e-06 2.060144e-06 0.2671969 6.680139e-06 2.500081e-05 3.2425588
#> 38  37 5.650067e-06 1.509680e-06 0.2671969 4.895227e-06 1.832067e-05 3.2425588
#> 39  38 4.140387e-06 1.106298e-06 0.2671969 3.587238e-06 1.342545e-05 3.2425588
#> 40  39 3.034088e-06 8.106989e-07 0.2671969 2.628739e-06 9.838210e-06 3.2425588
#> 41  40 2.223390e-06 5.940827e-07 0.2671969 1.926348e-06 7.209471e-06 3.2425588
#> 42  41 1.629307e-06 4.353457e-07 0.2671969 1.411634e-06 5.283123e-06 3.2425587
#> 43  42 1.193961e-06 3.190227e-07 0.2671969 1.034450e-06 3.871489e-06 3.2425587
#> 44  43 8.749385e-07 2.337808e-07 0.2671969 7.580480e-07 2.837039e-06 3.2425587
#> 45  44 6.411576e-07 1.713153e-07 0.2671969 5.555000e-07 2.078991e-06 3.2425587
#> 46  45 4.698423e-07 1.255404e-07 0.2671969 4.070721e-07 1.523491e-06 3.2425586
#> 47  46 3.443019e-07 9.199640e-08 0.2671969 2.983037e-07 1.116419e-06 3.2425586
#> 48  47 2.523055e-07 6.741525e-08 0.2671969 2.185979e-07 8.181154e-07 3.2425585
#> 49  48 1.848903e-07 4.940210e-08 0.2671969 1.601892e-07 5.995175e-07 3.2425584
#> 50  49 1.354882e-07 3.620202e-08 0.2671969 1.173872e-07 4.393283e-07 3.2425582
#> 51  50 9.928616e-08 2.652895e-08 0.2671969 8.602168e-08 3.219411e-07 3.2425580
#> 52  51 7.275721e-08 1.944050e-08 0.2671969 6.303696e-08 2.359194e-07 3.2425577
#> 53  52 5.331671e-08 1.424606e-08 0.2671969 4.619368e-08 1.728825e-07 3.2425573
#> 54  53 3.907065e-08 1.043956e-08 0.2671969 3.385087e-08 1.266888e-07 3.2425568
#> 55  54 2.863110e-08 7.650139e-09 0.2671969 2.480603e-08 9.283793e-08 3.2425561
#> 56  55 2.098096e-08 5.606046e-09 0.2671969 1.817793e-08 6.803191e-08 3.2425551
#> 57  56 1.537491e-08 4.108128e-09 0.2671969 1.332085e-08 4.985397e-08 3.2425537
#> 58  57 1.126678e-08 3.010449e-09 0.2671969 9.761558e-09 3.653313e-08 3.2425519
#> 59  58 8.256334e-09 2.206066e-09 0.2671969 7.153300e-09 2.677157e-08 3.2425493
#> 60  59 6.050267e-09 1.616612e-09 0.2671969 5.241961e-09 1.961827e-08 3.2425459
#> 61  60 4.433655e-09 1.184659e-09 0.2671969 3.841325e-09 1.437631e-08 3.2425412
#> 62  61 3.248996e-09 8.681216e-10 0.2671969 2.814935e-09 1.053498e-08 3.2425348
#> 63  62 2.380874e-09 6.361622e-10 0.2671969 2.062793e-09 7.720047e-09 3.2425260
#> 64  63 1.744712e-09 4.661816e-10 0.2671969 1.511621e-09 5.657254e-09 3.2425141
#> 65  64 1.278531e-09 3.416194e-10 0.2671969 1.107721e-09 4.145633e-09 3.2424978
#> 66  65 9.369112e-10 2.503397e-10 0.2671969 8.117413e-10 3.037912e-09 3.2424755
#> 67  66 6.865715e-10 1.834497e-10 0.2671969 5.948466e-10 2.226170e-09 3.2424452
#> 68  67 5.031217e-10 1.344325e-10 0.2671969 4.359054e-10 1.631324e-09 3.2424037
#> 69  68 3.686892e-10 9.851259e-11 0.2671969 3.194329e-10 1.195418e-09 3.2423472
#> 70  69 2.701766e-10 7.219034e-11 0.2671969 2.340814e-10 8.759854e-10 3.2422701
#> 71  70 1.979862e-10 5.290130e-11 0.2671969 1.715356e-10 6.419040e-10 3.2421648
#> 72  71 1.450849e-10 3.876624e-11 0.2671969 1.257018e-10 4.703684e-10 3.2420211
#> 73  72 1.063187e-10 2.840802e-11 0.2671969 9.211468e-11 3.446666e-10 3.2418251
#> 74  73 7.791067e-11 2.081749e-11 0.2671969 6.750193e-11 2.525519e-10 3.2415575
#> 75  74 5.709318e-11 1.525512e-11 0.2671969 4.946562e-11 1.850500e-10 3.2411924
#> 76  75 4.183806e-11 1.117900e-11 0.2671969 3.624856e-11 1.355844e-10 3.2406942
#> 77  76 3.065906e-11 8.192006e-12 0.2671969 2.656306e-11 9.933581e-11 3.2400143
#> 78  77 2.246706e-11 6.003128e-12 0.2671969 1.946549e-11 7.277275e-11 3.2390866
#> 79  78 1.646393e-11 4.399111e-12 0.2671969 1.426438e-11 5.330725e-11 3.2378205
#> 80  79 1.206482e-11 3.223682e-12 0.2671969 1.045298e-11 3.904288e-11 3.2360928
#> 81  80 8.841138e-12 2.362324e-12 0.2671969 7.659976e-12 2.858990e-11 3.2337352
#> 82  81 6.478813e-12 1.731119e-12 0.2671969 5.613254e-12 2.092992e-11 3.2305179
#> 83  82 4.747695e-12 1.268569e-12 0.2671969 4.113410e-12 1.531667e-11 3.2261275
#> 84  83 3.479126e-12 9.296115e-13 0.2671969 3.014320e-12 1.120326e-11 3.2201363
#> 85  84 2.549514e-12 6.812222e-13 0.2671969 2.208903e-12 8.188939e-12 3.2119605
#> 86  85 1.868292e-12 4.992018e-13 0.2671969 1.618691e-12 5.980036e-12 3.2008036
#> 87  86 1.369090e-12 3.658166e-13 0.2671969 1.186182e-12 4.361344e-12 3.1855787
#> 88  87 1.003274e-12 2.680716e-13 0.2671969 8.692378e-13 3.175163e-12 3.1648025
#> 89  88 7.352020e-13 1.964437e-13 0.2671969 6.369802e-13 2.305925e-12 3.1364507
#> 90  89 5.387583e-13 1.439545e-13 0.2671969 4.667811e-13 1.668945e-12 3.0977613
#> 91  90 3.948038e-13 1.054903e-13 0.2671969 3.420586e-13 1.202164e-12 3.0449648
#> 92  91 2.893134e-13 7.730365e-14 0.2671969 2.506616e-13 8.601050e-13 2.9729175
#> 93  92 2.120098e-13 5.664836e-14 0.2671969 1.836856e-13 6.094434e-13 2.8746001
#> 94  93 1.553614e-13 4.151209e-14 0.2671969 1.346054e-13 4.257578e-13 2.7404339
#> 95  94 1.138494e-13 3.042019e-14 0.2671969 9.863926e-14 2.911524e-13 2.5573476
#> 96  95 8.342916e-14 2.229201e-14 0.2671969 7.228316e-14 1.925131e-13 2.3075039
#> 97  96 6.113715e-14 1.633566e-14 0.2671969 5.296932e-14 1.202300e-13 1.9665614
#> 98  97 4.480150e-14 1.197082e-14 0.2671969 3.881609e-14 6.726064e-14 1.5013033
#> 99  98 3.283068e-14 8.772254e-15 0.2671969 2.844455e-14 2.844455e-14 0.8664016
#> 100 99 2.405842e-14           NA        NA           NA           NA        NA
#>            mx         lxmx        cx         lxcx     mxcx       lxmxcx
#> 1   0.0000000 0.000000e+00 0.0000000 0.000000e+00 0.000000 0.000000e+00
#> 2   0.0000000 0.000000e+00 0.0000000 0.000000e+00 0.000000 0.000000e+00
#> 3   0.9193548 2.850000e-01 0.3387097 1.050000e-01 1.258065 3.900000e-01
#> 4   1.4861111 3.210000e-01 0.5000000 1.080000e-01 1.986111 4.290000e-01
#> 5   1.7225016 2.740500e-01 0.5571967 8.865000e-02 2.279698 3.627000e-01
#> 6   1.8228528 2.139300e-01 0.5797546 6.804000e-02 2.402607 2.819700e-01
#> 7   1.8672533 1.611645e-01 0.5894787 5.087850e-02 2.456732 2.120430e-01
#> 8   1.8872827 1.195737e-01 0.5938293 3.762360e-02 2.481112 1.571973e-01
#> 9   1.8963856 8.811653e-02 0.5958016 2.768422e-02 2.492187 1.158007e-01
#> 10  1.9005343 6.473687e-02 0.5966998 2.032506e-02 2.497234 8.506192e-02
#> 11  1.9024270 4.749451e-02 0.5971095 1.490697e-02 2.499536 6.240149e-02
#> 12  1.9032909 3.482258e-02 0.5972964 1.092813e-02 2.500587 4.575070e-02
#> 13  1.9036853 2.552426e-02 0.5973818 8.009586e-03 2.501067 3.353385e-02
#> 14  1.9038653 1.870633e-02 0.5974208 5.869926e-03 2.501286 2.457625e-02
#> 15  1.9039475 1.370875e-02 0.5974386 4.301659e-03 2.501386 1.801040e-02
#> 16  1.9039850 1.004604e-02 0.5974467 3.152322e-03 2.501432 1.319836e-02
#> 17  1.9040022 7.361849e-03 0.5974504 2.310050e-03 2.501453 9.671898e-03
#> 18  1.9040100 5.394812e-03 0.5974521 1.692817e-03 2.501462 7.087629e-03
#> 19  1.9040136 3.953343e-03 0.5974528 1.240504e-03 2.501466 5.193847e-03
#> 20  1.9040152 2.897025e-03 0.5974532 9.090458e-04 2.501468 3.806071e-03
#> 21  1.9040159 2.122950e-03 0.5974534 6.661519e-04 2.501469 2.789102e-03
#> 22  1.9040163 1.555705e-03 0.5974534 4.881582e-04 2.501470 2.043863e-03
#> 23  1.9040164 1.140026e-03 0.5974535 3.577239e-04 2.501470 1.497749e-03
#> 24  1.9040165 8.354143e-04 0.5974535 2.621412e-04 2.501470 1.097556e-03
#> 25  1.9040165 6.121942e-04 0.5974535 1.920979e-04 2.501470 8.042921e-04
#> 26  1.9040165 4.486178e-04 0.5974535 1.407699e-04 2.501470 5.893878e-04
#> 27  1.9040166 3.287486e-04 0.5974535 1.031567e-04 2.501470 4.319052e-04
#> 28  1.9040166 2.409080e-04 0.5974535 7.559352e-05 2.501470 3.165015e-04
#> 29  1.9040166 1.765381e-04 0.5974535 5.539517e-05 2.501470 2.319333e-04
#> 30  1.9040166 1.293677e-04 0.5974535 4.059375e-05 2.501470 1.699614e-04
#> 31  1.9040166 9.480105e-05 0.5974535 2.974723e-05 2.501470 1.245483e-04
#> 32  1.9040166 6.947050e-05 0.5974535 2.179886e-05 2.501470 9.126937e-05
#> 33  1.9040166 5.090820e-05 0.5974535 1.597427e-05 2.501470 6.688248e-05
#> 34  1.9040166 3.730569e-05 0.5974535 1.170600e-05 2.501470 4.901169e-05
#> 35  1.9040166 2.733773e-05 0.5974535 8.578192e-06 2.501470 3.591592e-05
#> 36  1.9040166 2.003317e-05 0.5974535 6.286126e-06 2.501470 2.631930e-05
#> 37  1.9040166 1.468037e-05 0.5974535 4.606493e-06 2.501470 1.928686e-05
#> 38  1.9040166 1.075782e-05 0.5974535 3.375652e-06 2.501470 1.413347e-05
#> 39  1.9040166 7.883365e-06 0.5974535 2.473689e-06 2.501470 1.035705e-05
#> 40  1.9040166 5.776955e-06 0.5974535 1.812727e-06 2.501470 7.589681e-06
#> 41  1.9040166 4.233370e-06 0.5974535 1.328372e-06 2.501470 5.561742e-06
#> 42  1.9040166 3.102227e-06 0.5974535 9.734351e-07 2.501470 4.075662e-06
#> 43  1.9040166 2.273322e-06 0.5974535 7.133363e-07 2.501470 2.986658e-06
#> 44  1.9040166 1.665897e-06 0.5974535 5.227350e-07 2.501470 2.188632e-06
#> 45  1.9040166 1.220775e-06 0.5974535 3.830619e-07 2.501470 1.603837e-06
#> 46  1.9040166 8.945876e-07 0.5974535 2.807089e-07 2.501470 1.175296e-06
#> 47  1.9040166 6.555566e-07 0.5974535 2.057044e-07 2.501470 8.612609e-07
#> 48  1.9040166 4.803939e-07 0.5974535 1.507408e-07 2.501470 6.311347e-07
#> 49  1.9040166 3.520342e-07 0.5974535 1.104633e-07 2.501470 4.624975e-07
#> 50  1.9040166 2.579717e-07 0.5974535 8.094788e-08 2.501470 3.389196e-07
#> 51  1.9040166 1.890425e-07 0.5974535 5.931886e-08 2.501470 2.483614e-07
#> 52  1.9040166 1.385309e-07 0.5974535 4.346905e-08 2.501470 1.820000e-07
#> 53  1.9040166 1.015159e-07 0.5974535 3.185425e-08 2.501470 1.333702e-07
#> 54  1.9040166 7.439117e-08 0.5974535 2.334290e-08 2.501470 9.773406e-08
#> 55  1.9040166 5.451408e-08 0.5974535 1.710575e-08 2.501470 7.161983e-08
#> 56  1.9040166 3.994809e-08 0.5974535 1.253515e-08 2.501470 5.248323e-08
#> 57  1.9040166 2.927408e-08 0.5974535 9.185794e-09 2.501470 3.845988e-08
#> 58  1.9040166 2.145214e-08 0.5974535 6.731379e-09 2.501470 2.818352e-08
#> 59  1.9040166 1.572020e-08 0.5974535 4.932775e-09 2.501470 2.065297e-08
#> 60  1.9040166 1.151981e-08 0.5974535 3.614753e-09 2.501470 1.513456e-08
#> 61  1.9040166 8.441752e-09 0.5974535 2.648902e-09 2.501470 1.109065e-08
#> 62  1.9040166 6.186142e-09 0.5974535 1.941124e-09 2.501470 8.127266e-09
#> 63  1.9040166 4.533224e-09 0.5974535 1.422462e-09 2.501470 5.955686e-09
#> 64  1.9040166 3.321961e-09 0.5974535 1.042384e-09 2.501470 4.364345e-09
#> 65  1.9040166 2.434343e-09 0.5974535 7.638626e-10 2.501470 3.198206e-09
#> 66  1.9040166 1.783894e-09 0.5974535 5.597609e-10 2.501470 2.343655e-09
#> 67  1.9040166 1.307243e-09 0.5974535 4.101945e-10 2.501470 1.717438e-09
#> 68  1.9040166 9.579521e-10 0.5974535 3.005918e-10 2.501470 1.258544e-09
#> 69  1.9040166 7.019903e-10 0.5974535 2.202746e-10 2.501470 9.222649e-10
#> 70  1.9040166 5.144207e-10 0.5974535 1.614179e-10 2.501470 6.758386e-10
#> 71  1.9040166 3.769691e-10 0.5974535 1.182876e-10 2.501470 4.952567e-10
#> 72  1.9040166 2.762441e-10 0.5974535 8.668150e-11 2.501470 3.629256e-10
#> 73  1.9040166 2.024326e-10 0.5974535 6.352048e-11 2.501470 2.659530e-10
#> 74  1.9040166 1.483432e-10 0.5974535 4.654800e-11 2.501470 1.948912e-10
#> 75  1.9040166 1.087064e-10 0.5974535 3.411052e-11 2.501470 1.428169e-10
#> 76  1.9040166 7.966037e-11 0.5974535 2.499630e-11 2.501470 1.046567e-10
#> 77  1.9040166 5.837537e-11 0.5974535 1.831737e-11 2.501470 7.669273e-11
#> 78  1.9040166 4.277765e-11 0.5974535 1.342302e-11 2.501470 5.620067e-11
#> 79  1.9040166 3.134760e-11 0.5974535 9.836433e-12 2.501470 4.118403e-11
#> 80  1.9040166 2.297162e-11 0.5974535 7.208169e-12 2.501470 3.017979e-11
#> 81  1.9040166 1.683367e-11 0.5974535 5.282169e-12 2.501470 2.211584e-11
#> 82  1.9040166 1.233577e-11 0.5974535 3.870790e-12 2.501470 1.620656e-11
#> 83  1.9040166 9.039689e-12 0.5974535 2.836527e-12 2.501470 1.187622e-11
#> 84  1.9040166 6.624313e-12 0.5974535 2.078616e-12 2.501470 8.702928e-12
#> 85  1.9040166 4.854317e-12 0.5974535 1.523216e-12 2.501470 6.377533e-12
#> 86  1.9040166 3.557259e-12 0.5974535 1.116218e-12 2.501470 4.673476e-12
#> 87  1.9040166 2.606770e-12 0.5974535 8.179677e-13 2.501470 3.424738e-12
#> 88  1.9040166 1.910249e-12 0.5974535 5.994093e-13 2.501470 2.509659e-12
#> 89  1.9040166 1.399837e-12 0.5974535 4.392490e-13 2.501470 1.839086e-12
#> 90  1.9040166 1.025805e-12 0.5974535 3.218830e-13 2.501470 1.347688e-12
#> 91  1.9040166 7.517129e-13 0.5974535 2.358769e-13 2.501470 9.875898e-13
#> 92  1.9040166 5.508576e-13 0.5974535 1.728513e-13 2.501470 7.237089e-13
#> 93  1.9040166 4.036702e-13 0.5974535 1.266660e-13 2.501470 5.303362e-13
#> 94  1.9040166 2.958108e-13 0.5974535 9.282124e-14 2.501470 3.886320e-13
#> 95  1.9040166 2.167711e-13 0.5974535 6.801969e-14 2.501470 2.847907e-13
#> 96  1.9040166 1.588505e-13 0.5974535 4.984505e-14 2.501470 2.086956e-13
#> 97  1.9040166 1.164061e-13 0.5974535 3.652660e-14 2.501470 1.529328e-13
#> 98  1.9040166 8.530279e-14 0.5974535 2.676681e-14 2.501470 1.120696e-13
#> 99  1.9040166 6.251015e-14 0.5974535 1.961480e-14 2.501470 8.212495e-14
#> 100 1.9040166 4.580763e-14 0.5974535 1.437379e-14 2.501470 6.018142e-14