Perturbation analysis of matrix elements in a matrix population model

Perturbs elements within a matrix population model and measures the response of the per-capita population growth rate at equilibrium (\(\lambda\)), or, with a user-supplied function, any other demographic statistic.

matrixElementPerturbation(matU, matF, matC = NULL, pert = 1e-06,
  demogstat = "lambda", ...)

Arguments

matU
The survival component of a matrix population model (i.e. a square projection matrix reflecting survival-related transitions; e.g. progression, stasis, and retrogression)

matF
The sexual component of a matrix population model (i.e. a square projection matrix reflecting transitions due to sexual reproduction)

matC
The clonal component of a matrix population model (i.e. a square projection matrix reflecting transitions due to clonal reproduction). Defaults to NULL, indicating no clonal reproduction (i.e. matC is a matrix of zeros).

pert
Magnitude of the perturbation (defaults to 1e-6).

demogstat
The demographic statistic to be used, as in "the sensitivity/elasticity of ___ to matrix element perturbations." Defaults to the per-capita population growth rate at equilibrium (\(lambda\)). Also accepts a user-supplied function that performs a calculation on a projection matrix and returns a single numeric value.

...
Additional arguments passed to the function demogstat.

Value

A data frame with 1 row and 10 columns:

SStasis

sensitivity of demogstat to stasis

SRetrogression

sensitivity of demogstat to retrogression

SProgression

sensitivity of demogstat to progression

SFecundity

sensitivity of demogstat to sexual fecundity

SClonality

sensitivity of demogstat to clonality

EStasis

elasticity of demogstat to stasis

ERetrogression

elasticity of demogstat to retrogression

EProgression

elasticity of demogstat to progression

EFecundity

elasticity of demogstat to sexual fecundity

EClonality

elasticity of demogstat to clonality

Author

Rob Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Examples

matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

matrixElementPerturbation(matU, matF)

#>    SStasis SRetrogression SProgression SFecundity SClonality   EStasis
#> 1 1.000001      0.3211414     1.428538  0.2896157          0 0.2802673
#>   EProgression ERetrogression EFecundity EClonality
#> 1    0.4639941     0.03601974  0.2197187          0

# use a larger perturbation than the default
matrixElementPerturbation(matU, matF, pert = 0.01)

#>    SStasis SRetrogression SProgression SFecundity SClonality   EStasis
#> 1 1.005989      0.3206984     1.418785  0.2894348          0 0.2821539
#>   EProgression ERetrogression EFecundity EClonality
#> 1    0.4610285     0.03598399  0.2196051          0

# calculate the sensitivity/elasticity of the damping ratio to perturbations
damping <- function(matA) {  # define function for damping ratio
  eig <- eigen(matA)$values
  dm <- rle(Mod(eig))$values
  return(dm[1] / dm[2])
}

matrixElementPerturbation(matU, matF, demogstat = 'damping')

#>     SStasis SRetrogression SProgression SFecundity SClonality   EStasis
#> 1 0.2627447       2.493559    -4.831068   1.112042          0 0.6601846
#>   EProgression ERetrogression EFecundity EClonality
#> 1   -0.8706525      0.1230355 0.08743928          0