Generate a life table from a matrix population model

This function uses age-from-stage decomposition methods to generate a life table from a matrix population model. A detailed description of these methods can be found in section 5.3 "Age-specific traits from stage-specific models" of Caswell (2001).

mpm_to_table(
  matU,
  matF = NULL,
  matC = NULL,
  start = 1L,
  xmax = 1000,
  lx_crit = 1e-04,
  radix = 1
)

Arguments

matU
The survival component of a matrix population model (i.e. a square projection matrix reflecting survival-related transitions; e.g. progression, stasis, and/or retrogression). Optionally with named rows and columns indicating the corresponding life stage names.

matF
(Optional) The sexual component of a matrix population model (i.e. a square projection matrix reflecting transitions due to sexual reproduction). Optionally with named rows and columns indicating the corresponding life stage names.

matC
(Optional) The clonal component of a matrix population model (i.e. a square projection matrix reflecting transitions due to clonal reproduction). Optionally with named rows and columns indicating the corresponding life stage names.

start
The index (or stage name) of the first stage at which the author considers the beginning of life. Defaults to 1. Alternately, a numeric vector giving the starting population vector (in which case length(start) must match ncol(matU)). See section Starting from multiple stages.

xmax
Maximum age to which the life table will be calculated (defaults to 1000). Time steps are in the same units as the matrix population model (see MatrixPeriodicity metadata variable COM(P)ADRE).

lx_crit
Minimum value of lx to which age-specific traits will be calculated (defaults to 1e-4).

radix
The starting number of individuals in the synthetic life table (defaults to 1). If radix is set to 1, a simplified life table is produced.

Value

A data.frame containing a variable number columns, depending on input variables. Columns include:

x

age at the start of the age interval [x, x+1)

Nx

The number of individuals alive at age x. The initial number is set with radix

Dx

proportion of original cohort dying during the age interval [x, x+1)

lx

survivorship, defined as the proportion of initial cohort surviving to the start of age interval [x, x+1)

dx

proportion of original cohort dying in the age interval [x, x+1)

ax

The average time survived within the interval by those that die during the age interval [x, x+1). Assumed to be 0.5

hx

force of mortality (hazard) during the age interval [x, x+1)

qx

probability of death during the interval [x, x+1) for those entering the interval

px

probability of survival for the interval [x, x+1) for those entering the interval

Lx

total person-years lived during the interval [x, x+1)

Tx

total person years lived beyond age x

ex

remaining life expectancy at age x

If matF is provided, also includes:
mx

per-capita rate of sexual reproduction during the interval [x, x+1)

lxmx

expected number of sexual offspring per original cohort member produced during the interval [x, x+1)

If matC is provided, also includes:
cx

per-capita rate of clonal reproduction during the interval [x, x+1)

lxcx

expected number of clonal offspring per original cohort member produced during the interval [x, x+1)

If both matF and matC are provided, also includes:
mxcx

per-capita rate of total reproduction (sexual + clonal) during the interval [x, x+1)

lxmxcx

expected number of total offspring (sexual + clonal) per original cohort member produced during the interval [x, x+1)

Note

The life table is calculated recursively until the age class (x) reaches xmax or survivorship (lx) falls below lx_crit — whichever comes first. To force calculation to xmax, set lx_crit = 0. Conversely, to force calculation to lx_crit, set xmax = Inf.

The life table calculations assume that the final age interval is closed and that all remaining individuals die in this interval. Therefore, for this interval, the probability of death qx is 1, the probability of survival px is 0 and, because we assume that deaths are evenly distributed during the interval, the remaining life expectancy for individuals at the start of the interval is 0.5.

If lx_crit is sufficiently small that only a very small proportion of the cohort reach this age (i.e. < 0.05), this should have minimal impact on results. Nevertheless, for many analyses, the final row of the life table should be treated with caution and perhaps removed from subsequent analyses.

Starting from multiple stages

Rather than specifying argument start as a single stage class from which all individuals start life, it may sometimes be desirable to allow for multiple starting stage classes. For example, if the user wants to start the calculation of age-specific traits from reproductive maturity (i.e. first reproduction), the user should account for the possibility that there may be multiple stage classes in which an individual could first reproduce.

To specify multiple starting stage classes, specify argument start as the desired starting population vector (n1), giving the proportion of individuals starting in each stage class (the length of start should match the number of columns in the relevant MPM).

See function mature_distrib for calculating the proportion of individuals achieving reproductive maturity in each stage class.

References

Caswell, H. 2001. Matrix Population Models: Construction, Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

Caswell, H. 2006. Applications of Markov chains in demography. pp. 319-334 in A.N. Langville and W.J. Stewart (editors) MAM2006: Markov Anniversary Meeting. Boson Books, Raleigh, North Caroline, USA

Horvitz, C. & Tuljapurkar, S. 2008. Stage dynamics, period survival, and mortality plateaus. The American Naturalist 172: 203-2015. <doi:10.1086/589453>

Jones, O. R., Scheuerlein, A., Salguero-Gomez, R., Camarda, C. G., Schaible, R., Casper, B. B., Dahlgren, J. P., Ehrlén, J., García, M. B., Menges, E., Quintana-Ascencio, P. F., Caswell, H., Baudisch, A. & Vaupel, J. 2014. Diversity of ageing across the tree of life. Nature 505, 169-173. <doi:10.1038/nature12789>

Jones O. R. 2021. Life tables: Construction and interpretation In: Demographic Methods Across the Tree of Life. Edited by Salguero-Gomez R & Gamelon M. Oxford University Press. Oxford, UK. ISBN: 9780198838609

Preston, S., Heuveline, P., & Guillot, M. 2000. Demography: Measuring and Modeling Population Processes. Wiley. ISBN: 9781557864512

See also

Author

Owen R. Jones <jones@biology.sdu.dk>

Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Examples

data(mpm1)

mpm_to_table(matU = mpm1$matU, start = 2, xmax = 15)

#>     x          lx          dx        hx        qx        px       ex
#> 1   0 1.000000000 0.500000000 0.6666667 0.5000000 0.5000000 2.001896
#> 2   1 0.500000000 0.183200000 0.4485798 0.3664000 0.6336000 2.503791
#> 3   2 0.316800000 0.108187000 0.4118170 0.3414994 0.6585006 2.662549
#> 4   3 0.208613000 0.064699290 0.3670604 0.3101403 0.6898597 2.784051
#> 5   4 0.143913710 0.043329919 0.3544406 0.3010826 0.6989174 2.810891
#> 6   5 0.100583791 0.029887091 0.3489842 0.2971363 0.7028637 2.806387
#> 7   6 0.070696700 0.020921617 0.3473281 0.2959348 0.7040652 2.781414
#> 8   7 0.049775083 0.014708992 0.3467418 0.2955091 0.7044909 2.740345
#> 9   8 0.035066091 0.010357661 0.3465577 0.2953754 0.7046246 2.680091
#> 10  9 0.024708430 0.007297162 0.3464964 0.2953309 0.7046691 2.593975
#> 11 10 0.017411268 0.005141838 0.3464769 0.2953167 0.7046833 2.471571
#> 12 11 0.012269430 0.003623311 0.3464705 0.2953121 0.7046879 2.297812
#> 13 12 0.008646119 0.002553291 0.3464685 0.2953106 0.7046894 2.051217
#> 14 13 0.006092829 0.001799274 0.3464678 0.2953101 0.7046899 1.701278
#> 15 14 0.004293555 0.001267929 0.3464676 0.2953100 0.7046900 1.204690
#> 16 15 0.003025625 0.003025625 2.0000000 1.0000000 0.0000000 0.500000

mpm_to_table(matU = mpm1$matU, start = "small", xmax = 15) # equivalent using named life stages

#>     x          lx          dx        hx        qx        px       ex
#> 1   0 1.000000000 0.500000000 0.6666667 0.5000000 0.5000000 2.001896
#> 2   1 0.500000000 0.183200000 0.4485798 0.3664000 0.6336000 2.503791
#> 3   2 0.316800000 0.108187000 0.4118170 0.3414994 0.6585006 2.662549
#> 4   3 0.208613000 0.064699290 0.3670604 0.3101403 0.6898597 2.784051
#> 5   4 0.143913710 0.043329919 0.3544406 0.3010826 0.6989174 2.810891
#> 6   5 0.100583791 0.029887091 0.3489842 0.2971363 0.7028637 2.806387
#> 7   6 0.070696700 0.020921617 0.3473281 0.2959348 0.7040652 2.781414
#> 8   7 0.049775083 0.014708992 0.3467418 0.2955091 0.7044909 2.740345
#> 9   8 0.035066091 0.010357661 0.3465577 0.2953754 0.7046246 2.680091
#> 10  9 0.024708430 0.007297162 0.3464964 0.2953309 0.7046691 2.593975
#> 11 10 0.017411268 0.005141838 0.3464769 0.2953167 0.7046833 2.471571
#> 12 11 0.012269430 0.003623311 0.3464705 0.2953121 0.7046879 2.297812
#> 13 12 0.008646119 0.002553291 0.3464685 0.2953106 0.7046894 2.051217
#> 14 13 0.006092829 0.001799274 0.3464678 0.2953101 0.7046899 1.701278
#> 15 14 0.004293555 0.001267929 0.3464676 0.2953100 0.7046900 1.204690
#> 16 15 0.003025625 0.003025625 2.0000000 1.0000000 0.0000000 0.500000

mpm_to_table(matU = mpm1$matU, matF = mpm1$matF, start = 2, xmax = 15)

#>     x          lx          dx        hx        qx        px       ex       mx
#> 1   0 1.000000000 0.500000000 0.6666667 0.5000000 0.5000000 2.001896  0.00000
#> 2   1 0.500000000 0.183200000 0.4485798 0.3664000 0.6336000 2.503791 15.26600
#> 3   2 0.316800000 0.108187000 0.4118170 0.3414994 0.6585006 2.662549 22.00578
#> 4   3 0.208613000 0.064699290 0.3670604 0.3101403 0.6898597 2.784051 25.84823
#> 5   4 0.143913710 0.043329919 0.3544406 0.3010826 0.6989174 2.810891 27.13917
#> 6   5 0.100583791 0.029887091 0.3489842 0.2971363 0.7028637 2.806387 27.62496
#> 7   6 0.070696700 0.020921617 0.3473281 0.2959348 0.7040652 2.781414 27.78025
#> 8   7 0.049775083 0.014708992 0.3467418 0.2955091 0.7044909 2.740345 27.83270
#> 9   8 0.035066091 0.010357661 0.3465577 0.2953754 0.7046246 2.680091 27.84948
#> 10  9 0.024708430 0.007297162 0.3464964 0.2953309 0.7046691 2.593975 27.85497
#> 11 10 0.017411268 0.005141838 0.3464769 0.2953167 0.7046833 2.471571 27.85673
#> 12 11 0.012269430 0.003623311 0.3464705 0.2953121 0.7046879 2.297812 27.85730
#> 13 12 0.008646119 0.002553291 0.3464685 0.2953106 0.7046894 2.051217 27.85749
#> 14 13 0.006092829 0.001799274 0.3464678 0.2953101 0.7046899 1.701278 27.85755
#> 15 14 0.004293555 0.001267929 0.3464676 0.2953100 0.7046900 1.204690 27.85756
#> 16 15 0.003025625 0.003025625 2.0000000 1.0000000 0.0000000 0.500000 27.85757
#>          lxmx
#> 1  0.00000000
#> 2  7.63300000
#> 3  6.97143000
#> 4  5.39227780
#> 5  3.90569858
#> 6  2.77862313
#> 7  1.96397169
#> 8  1.38537517
#> 9  0.97657248
#> 10 0.68825265
#> 11 0.48502105
#> 12 0.34179324
#> 13 0.24085915
#> 14 0.16973126
#> 15 0.11960798
#> 16 0.08428657

### starting from first reproduction
repStages <- repro_stages(mpm1$matF)
n1 <- mature_distrib(matU = mpm1$matU, start = 2, repro_stages = repStages)
mpm_to_table(matU = mpm1$matU, start = n1)

#>     x           lx           dx        hx        qx        px       ex
#> 1   0 1.000000e+00 3.242105e-01 0.3869347 0.3242105 0.6757895 2.678768
#> 2   1 6.757895e-01 2.268500e-01 0.4033861 0.3356815 0.6643185 2.724034
#> 3   2 4.489395e-01 1.360970e-01 0.3573121 0.3031522 0.6968478 2.847843
#> 4   3 3.128425e-01 9.359475e-02 0.3518003 0.2991753 0.7008247 2.869233
#> 5   4 2.192478e-01 6.496711e-02 0.3478563 0.2963182 0.7036818 2.880635
#> 6   5 1.542806e-01 4.561886e-02 0.3469874 0.2956875 0.7043125 2.883114
#> 7   6 1.086618e-01 3.210105e-02 0.3466214 0.2954217 0.7045783 2.883603
#> 8   7 7.656074e-02 2.261206e-02 0.3465199 0.2953479 0.7046521 2.883020
#> 9   8 5.394868e-02 1.593222e-02 0.3464838 0.2953217 0.7046783 2.881839
#> 10  9 3.801646e-02 1.122679e-02 0.3464729 0.2953138 0.7046862 2.880038
#> 11 10 2.678968e-02 7.911290e-03 0.3464692 0.2953111 0.7046889 2.877444
#> 12 11 1.887839e-02 5.574982e-03 0.3464681 0.2953103 0.7046897 2.873749
#> 13 12 1.330341e-02 3.928629e-03 0.3464677 0.2953100 0.7046900 2.868503
#> 14 13 9.374777e-03 2.768465e-03 0.3464676 0.2953099 0.7046901 2.861056
#> 15 14 6.606313e-03 1.950909e-03 0.3464675 0.2953099 0.7046901 2.850489
#> 16 15 4.655403e-03 1.374787e-03 0.3464675 0.2953099 0.7046901 2.835493
#> 17 16 3.280617e-03 9.687985e-04 0.3464675 0.2953099 0.7046901 2.814212
#> 18 17 2.311818e-03 6.827027e-04 0.3464675 0.2953099 0.7046901 2.784014
#> 19 18 1.629115e-03 4.810939e-04 0.3464675 0.2953099 0.7046901 2.741161
#> 20 19 1.148022e-03 3.390221e-04 0.3464675 0.2953099 0.7046901 2.680349
#> 21 20 8.089994e-04 2.389055e-04 0.3464675 0.2953099 0.7046901 2.594054
#> 22 21 5.700939e-04 1.683544e-04 0.3464675 0.2953099 0.7046901 2.471595
#> 23 22 4.017395e-04 1.186377e-04 0.3464675 0.2953099 0.7046901 2.297819
#> 24 23 2.831019e-04 8.360278e-05 0.3464675 0.2953099 0.7046901 2.051219
#> 25 24 1.994991e-04 5.891406e-05 0.3464675 0.2953099 0.7046901 1.701278
#> 26 25 1.405850e-04 4.151615e-05 0.3464675 0.2953099 0.7046901 1.204690
#> 27 26 9.906889e-05 9.906889e-05 2.0000000 1.0000000 0.0000000 0.500000