Calculate stochastic elasticities from a time-series of matrix population models and corresponding population vectors

Calculate stochastic elasticities given a time-series of matrix population models and corresponding population vectors, using the method described in Haridas et al. (2009).

perturb_stochastic(X_t, u_t)

Arguments

X_t

A list of matrix population models

u_t

A list of corresponding population vectors

Value

A list of three matrices:

E

matrix of stochastic elasticities

E_mu

matrix of stochastic elasticities to mean transition rates

E_sigma

matrix of stochastic elasticities to the variance in transition rates

References

Haridas, C. V., Tuljapurkar, S., & Coulson, T. 2009. Estimating stochastic elasticities directly from longitudinal data. Ecology Letters, 12, 806-812. <doi:10.1111/j.1461-0248.2009.01330.x>

See also

Author

Patrick Barks <patrick.barks@gmail.com>

Examples

# generate list of random MPMs
N <- 20     # number of years
s <- 3      # matrix dimension
X <- list() # matrix population model at time t
u <- list() # population vector at time t

for(t in 1:N) {
  X[[t]] <- matrix(runif(s^2), nrow = s, ncol = s)
}

# derive corresponding series of population vectors
u <- pop_vectors(X)

# calculate stochastic elasticities
perturb_stochastic(X, u)
#> $E
#>           [,1]      [,2]       [,3]
#> [1,] 0.1129276 0.1334631 0.11523323
#> [2,] 0.1324792 0.1067274 0.10365725
#> [3,] 0.1120954 0.1007735 0.08264331
#> 
#> $E_mu
#>           [,1]      [,2]       [,3]
#> [1,] 0.1111123 0.1345361 0.12065276
#> [2,] 0.1321984 0.1083170 0.11874568
#> [3,] 0.1118820 0.1012017 0.09770496
#> 
#> $E_sigma
#>              [,1]          [,2]         [,3]
#> [1,] 0.0018153448 -0.0010729720 -0.005419539
#> [2,] 0.0002807500 -0.0015895736 -0.015088427
#> [3,] 0.0002134341 -0.0004281479 -0.015061651
#>