Derive a hypothetical set of population vectors corresponding to a time-series of matrix population models

Derive a hypothetical set of population vectors (i.e. population size distributions across stages) given a time-series of matrix population models (MPMs), by taking the stable stage distribution of the mean matrix as the starting vector (or optionally, a uniform or random starting vector), and deriving subsequent vectors through recursive population projection.

pop_vectors(A, start = "stable.stage")

Arguments

A

A list of MPMs (i.e., square population projection matrices).

start

Method to derive the first population vector in the series. Either stable.stage to use the stable stage distribution of the mean matrix as the starting vector, uniform to use a uniform starting vector (all elements equal), or random to use a randomly-generated starting vector. Defaults to the stable stage distribution.

Value

A list of population vectors

Details

This function is useful for providing population vectors as input to the perturb_stochastic function which calculates stochastic elasticities given a time-series of matrix population models and corresponding population vectors, using the method described in Haridas et al. (2009).

References

Haridas, C. V., Tuljapurkar, S., & Coulson, T. 2009. Estimating stochastic elasticities directly from longitudinal data. Ecology Letters, 12, 806-812. <doi:10.1111/j.1461-0248.2009.01330.x>

See also

Author

Patrick Barks <patrick.barks@gmail.com>

Examples

# generate list of matrices
matA_l <- replicate(5, matrix(runif(9), 3, 3), simplify = FALSE)

# calculate corresponding population vectors
pop_vectors(matA_l)
#> [[1]]
#> [1] 0.3369236 0.3614924 0.3015840
#> 
#> [[2]]
#> [1] 0.3726470 0.4108725 0.2164805
#> 
#> [[3]]
#> [1] 0.2526338 0.2779577 0.4694084
#> 
#> [[4]]
#> [1] 0.4768183 0.2863179 0.2368638
#> 
#> [[5]]
#> [1] 0.1212235 0.4591106 0.4196659
#> 
pop_vectors(matA_l, start = "uniform")
#> [[1]]
#> [1] 0.3333333 0.3333333 0.3333333
#> 
#> [[2]]
#> [1] 0.3866487 0.4037359 0.2096154
#> 
#> [[3]]
#> [1] 0.2519486 0.2744115 0.4736400
#> 
#> [[4]]
#> [1] 0.4760908 0.2869450 0.2369642
#> 
#> [[5]]
#> [1] 0.1213718 0.4590586 0.4195696
#> 
pop_vectors(matA_l, start = "random")
#> [[1]]
#> [1] 0.19241936 0.76054873 0.04703191
#> 
#> [[2]]
#> [1] 0.2401312 0.4440006 0.3158682
#> 
#> [[3]]
#> [1] 0.2656467 0.3066279 0.4277254
#> 
#> [[4]]
#> [1] 0.4854946 0.2806050 0.2339004
#> 
#> [[5]]
#> [1] 0.1194373 0.4603354 0.4202273
#>