Calculate age-specific traits from a matrix population model

These functions use age-from-stage decomposition methods to calculate age-specific survivorship (lx), survival probability (px), mortality hazard (hx), or reproduction (mx) from a matrix population model (MPM). A detailed description of these methods can be found in sections 5.3.1 and 5.3.2 of Caswell (2001).

mpm_to_mx(matU, matR, start = 1L, xmax = 1e+05, lx_crit = 1e-04, tol = 1e-04)

mpm_to_lx(matU, start = 1L, xmax = 1e+05, lx_crit = 1e-04, tol = 1e-04)

mpm_to_px(matU, start = 1L, xmax = 1e+05, lx_crit = 1e-04, tol = 1e-04)

mpm_to_hx(matU, start = 1L, xmax = 1e+05, lx_crit = 1e-04, tol = 1e-04)

Arguments

matU
The survival component of a MPM (i.e. a square projection matrix reflecting survival-related transitions; e.g. progression, stasis, and retrogression). Optionally with named rows and columns indicating the corresponding life stage names.

matR
The reproductive component of a MPM (i.e. a square projection matrix reflecting transitions due to reproduction; either sexual, clonal, or both). Optionally with named rows and columns indicating the corresponding life stage names.

start
The index (or stage name) of the first stage at which the author considers the beginning of life. Defaults to 1. Alternately, a numeric vector giving the starting population vector (in which case length(start) must match ncol(matU)). See section Starting from multiple stages.

xmax
Maximum age to which age-specific traits will be calculated (defaults to 100000).

lx_crit
Minimum value of lx to which age-specific traits will be calculated (defaults to 0.0001).

tol
To account for floating point errors that occasionally lead to values of lx slightly greater than 1, values of lx within the open interval (1, 1 + tol) are coerced to 1. Defaults to 0.0001. To prevent coercion, set tol to 0.

Value

A vector

Note

The output vector is calculated recursively until the age class (x) reaches xmax or survivorship (lx) falls below lx_crit, whichever comes first. To force calculation to xmax, set lx_crit to 0. Conversely, to force calculation to lx_crit, set xmax to Inf.

Starting from multiple stages

Rather than specifying argument start as a single stage class from which all individuals start life, it may sometimes be desirable to allow for multiple starting stage classes. For example, if users want to start their calculation of age-specific traits from reproductive maturity (i.e. first reproduction), they should account for the possibility that there may be multiple stage classes in which an individual could first reproduce.

To specify multiple starting stage classes, users should specify argument start as the desired starting population vector (n1), giving the proportion of individuals starting in each stage class (the length of start should match the number of columns in the relevant MPM).

See function mature_distrib for calculating the proportion of individuals achieving reproductive maturity in each stage class.

References

Caswell, H. 2001. Matrix Population Models: Construction, Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

See also

Author

Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>

Patrick Barks <patrick.barks@gmail.com>

Examples

data(mpm1)

# age-specific survivorship
mpm_to_lx(mpm1$matU)

#>  [1] 1.000000e+00 1.500000e-01 4.000000e-02 1.984000e-02 1.241465e-02
#>  [6] 8.437150e-03 5.872905e-03 4.122125e-03 2.900967e-03 2.043401e-03
#> [11] 1.439762e-03 1.014540e-03 7.149255e-04 5.037985e-04 3.550213e-04
#> [16] 2.501799e-04 1.762993e-04 1.242363e-04 8.754812e-05

mpm_to_lx(mpm1$matU, start = 2)       # starting from stage 2

#>  [1] 1.0000000000 0.5000000000 0.3168000000 0.2086130000 0.1439137100
#>  [6] 0.1005837911 0.0706967003 0.0497750832 0.0350660912 0.0247084304
#> [11] 0.0174112681 0.0122694301 0.0086461194 0.0060928288 0.0042935549
#> [16] 0.0030256254 0.0021321283 0.0015024897 0.0010587897 0.0007461186
#> [21] 0.0005257824 0.0003705137 0.0002610973 0.0001839927 0.0001296578
#> [26] 0.0000913686

mpm_to_lx(mpm1$matU, start = "small") # equivalent using named life stages

#>  [1] 1.0000000000 0.5000000000 0.3168000000 0.2086130000 0.1439137100
#>  [6] 0.1005837911 0.0706967003 0.0497750832 0.0350660912 0.0247084304
#> [11] 0.0174112681 0.0122694301 0.0086461194 0.0060928288 0.0042935549
#> [16] 0.0030256254 0.0021321283 0.0015024897 0.0010587897 0.0007461186
#> [21] 0.0005257824 0.0003705137 0.0002610973 0.0001839927 0.0001296578
#> [26] 0.0000913686

mpm_to_lx(mpm1$matU, xmax = 10)       # to a maximum age of 10

#>  [1] 1.000000000 0.150000000 0.040000000 0.019840000 0.012414650 0.008437150
#>  [7] 0.005872905 0.004122125 0.002900967 0.002043401 0.001439762

mpm_to_lx(mpm1$matU, lx_crit = 0.05)  # to a minimum lx of 0.05

#> [1] 1.00 0.15 0.04

# age-specific survival probability
mpm_to_px(mpm1$matU)

#>  [1] 0.1500000 0.2666667 0.4960000 0.6257384 0.6796124 0.6960768 0.7018887
#>  [8] 0.7037551 0.7043863 0.7045908 0.7046580 0.7046797 0.7046868 0.7046890
#> [15] 0.7046898 0.7046900 0.7046901 0.7046901        NA

# age-specific mortality hazard
mpm_to_hx(mpm1$matU)

#>  [1] 1.8971200 1.3217558 0.7011794 0.4688229 0.3862326 0.3622953 0.3539804
#>  [8] 0.3513249 0.3504283 0.3501381 0.3500427 0.3500119 0.3500019 0.3499987
#> [15] 0.3499976 0.3499973 0.3499972 0.3499971        NA

# age-specific fecundity
mpm_to_mx(mpm1$matU, mpm1$matF)

#>  [1]  0.00000  0.00000  9.54125 19.49277 24.83256 26.79977 27.50641 27.74123
#>  [9] 27.81971 27.84525 27.85359 27.85629 27.85716 27.85744 27.85753 27.85756
#> [17] 27.85757 27.85757 27.85757

### starting from first reproduction
repstages <- repro_stages(mpm1$matF)
n1 <- mature_distrib(mpm1$matU, start = 2, repro_stages = repstages)

mpm_to_lx(mpm1$matU, start = n1)

#>  [1] 1.000000e+00 6.757895e-01 4.489395e-01 3.128425e-01 2.192478e-01
#>  [6] 1.542806e-01 1.086618e-01 7.656074e-02 5.394868e-02 3.801646e-02
#> [11] 2.678968e-02 1.887839e-02 1.330341e-02 9.374777e-03 6.606313e-03
#> [16] 4.655403e-03 3.280617e-03 2.311818e-03 1.629115e-03 1.148022e-03
#> [21] 8.089994e-04 5.700939e-04 4.017395e-04 2.831019e-04 1.994991e-04
#> [26] 1.405850e-04 9.906889e-05

mpm_to_px(mpm1$matU, start = n1)

#>  [1] 0.6757895 0.6643185 0.6968478 0.7008247 0.7036818 0.7043125 0.7045783
#>  [8] 0.7046521 0.7046783 0.7046862 0.7046889 0.7046897 0.7046900 0.7046901
#> [15] 0.7046901 0.7046901 0.7046901 0.7046901 0.7046901 0.7046901 0.7046901
#> [22] 0.7046901 0.7046901 0.7046901 0.7046901 0.7046901        NA

mpm_to_hx(mpm1$matU, start = n1)

#>  [1] 0.3918737 0.4089935 0.3611882 0.3554975 0.3514291 0.3505331 0.3501558
#>  [8] 0.3500511 0.3500139 0.3500026 0.3499989 0.3499977 0.3499973 0.3499972
#> [15] 0.3499971 0.3499971 0.3499971 0.3499971 0.3499971 0.3499971 0.3499971
#> [22] 0.3499971 0.3499971 0.3499971 0.3499971 0.3499971        NA

mpm_to_mx(mpm1$matU, mpm1$matF, start = n1)

#>  [1] 20.08684 23.58049 26.70449 27.41101 27.72566 27.81222 27.84348 27.85293
#>  [9] 27.85610 27.85709 27.85742 27.85752 27.85756 27.85757 27.85757 27.85757
#> [17] 27.85757 27.85757 27.85757 27.85757 27.85757 27.85757 27.85757 27.85757
#> [25] 27.85757 27.85757 27.85757