Calculate mean and variance of life expectancy from a matrix population model

Applies Markov chain approaches to obtain mean and variance of life expectancy from a matrix population model (MPM).

life_expect_mean(matU, start = 1L)

life_expect_var(matU, start = 1L)

life_expect(matU, start = 1L)

Arguments

matU

The survival component of a MPM (i.e., a square projection matrix reflecting survival-related transitions; e.g., progression, stasis, and retrogression). Optionally with named rows and columns indicating the corresponding life stage names.

start

The index (or stage name) of the first stage of the life cycle which the user considers to be the beginning of life. Defaults to 1. Alternately, a numeric vector giving the starting population vector (in which case length(start) must match ncol(matU)). See section Starting from multiple stages.

Value

Returns life expectancy. If matU is singular (often indicating infinite life expectancy), returns NA.

Note

Note that the units of time in returned values are the same as the projection interval (`ProjectionInterval`) of the MPM.

Starting from multiple stages

Rather than specifying argument start as a single stage class from which all individuals start life, it may sometimes be desirable to allow for multiple starting stage classes. For example, if the user wants to start their calculation of life expectancy from reproductive maturity (i.e., first reproduction), they should account for the possibility that there may be multiple stage classes in which an individual could first reproduce.

To specify multiple starting stage classes, specify argument start as the desired starting population vector (n1), giving the proportion of individuals starting in each stage class (the length of start should match the number of columns in the relevant MPM).

See function mature_distrib for calculating the proportion of individuals achieving reproductive maturity in each stage class.

References

Caswell, H. 2001. Matrix Population Models: Construction, Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

See also

Author

Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <hcaswell@whoi.edu>

Examples

data(mpm1)

# mean life expectancy starting from stage class 2 
life_expect_mean(mpm1$matU, start = 2)
#> [1] 2.509116
life_expect_mean(mpm1$matU, start = "small")  # equivalent using named life stages
#> [1] 2.509116

# mean life expectancy starting from first reproduction
rep_stages <- repro_stages(mpm1$matF)
n1 <- mature_distrib(mpm1$matU, start = 2, repro_stages = rep_stages)
life_expect_mean(mpm1$matU, start = n1)
#> [1] 3.179005

# variance of life expectancy from stage class 1
life_expect_var(mpm1$matU, start = 1)
#> [1] 19.23639