Convert matrix population model into U, F and C matrices

Splits a matrix population model into three constituent matrices, U (growth and survival processes), F (sexual reproduction) and C (clonal reproduction). Warning! The functionality is very basic: it assumes that sexual reproduction is located in the top row of the matrix, and that everything else is growth or survival (i.e. the U matrix). Clonality is assumed to be non-existent.

mpm_split(matA)

Arguments

matA

A matrix population model (i.e., a square projection matrix).

Value

A list of three matrices: matU,matF and matC. matC will always contain only zeros.

See also

Author

Owen R. Jones <jones@biology.sdu.dk>

Examples

matA <- rbind(
  c(0.1, 0, 5.3, 4.2),
  c(0.5, 0.2, 0.1, 0),
  c(0, 0.3, 0.3, 0.1),
  c(0, 0, 0.5, 0.6)
)

mpm_split(matA)
#> $matU
#>      [,1] [,2] [,3] [,4]
#> [1,]  0.0  0.0  0.0  0.0
#> [2,]  0.5  0.2  0.1  0.0
#> [3,]  0.0  0.3  0.3  0.1
#> [4,]  0.0  0.0  0.5  0.6
#> 
#> $matF
#>      [,1] [,2] [,3] [,4]
#> [1,]  0.1    0  5.3  4.2
#> [2,]  0.0    0  0.0  0.0
#> [3,]  0.0    0  0.0  0.0
#> [4,]  0.0    0  0.0  0.0
#> 
#> $matC
#>      [,1] [,2] [,3] [,4]
#> [1,]    0    0    0    0
#> [2,]    0    0    0    0
#> [3,]    0    0    0    0
#> [4,]    0    0    0    0
#>