Generate a life table from a matrix population model

This function uses age-from-stage decomposition methods to generate a life table from a matrix population model. A detailed description of these methods can be found in section 5.3 "Age-specific traits from stage-specific models" of Caswell (2001).

mpm_to_table(
  matU,
  matF = NULL,
  matC = NULL,
  start = 1L,
  xmax = 1000,
  lx_crit = 1e-04
)

Arguments

matU
The survival component of a matrix population model (i.e. a square projection matrix reflecting survival-related transitions; e.g. progression, stasis, and/or retrogression). Optionally with named rows and columns indicating the corresponding life stage names.

matF
(Optional) The sexual component of a matrix population model (i.e. a square projection matrix reflecting transitions due to sexual reproduction). Optionally with named rows and columns indicating the corresponding life stage names.

matC
(Optional) The clonal component of a matrix population model (i.e. a square projection matrix reflecting transitions due to clonal reproduction). Optionally with named rows and columns indicating the corresponding life stage names.

start
The index (or stage name) of the first stage at which the author considers the beginning of life. Defaults to 1. Alternately, a numeric vector giving the starting population vector (in which case length(start) must match ncol(matU)). See section Starting from multiple stages.

xmax
Maximum age to which the life table will be calculated (defaults to 1000). Time steps are in the same units as the matrix population model (see MatrixPeriodicity metadata variable COM(P)ADRE).

lx_crit
Minimum value of lx to which age-specific traits will be calculated (defaults to 1e-4).

Value

A data.frame containing 7-13 columns, depending on input variables. Columns include:

x

age.

lx

survivorship to start of age class x.

dx

proportion of original cohort dying in interval [x, x+1).

qx

force of mortality at age x.

Lx

survivorship to middle of age class x.

Tx

proportion of original cohort alive at age x and beyond.

ex

remaining life expectancy at age x.

If matF is provided, also includes:
mx

per-capita rate of sexual reproduction at age x.

lxmx

expected number of sexual offspring per original cohort member produced at age x.

If matC is provided, also includes:
cx

per-capita rate of clonal reproduction at age x.

lxcx

expected number of clonal offspring per original cohort member produced at age x.

If both matF and matC are provided, also includes:
mxcx

per-capita rate of total reproduction (sexual + clonal) at age x.

lxmxcx

expected number of total offspring (sexual + clonal) per original cohort member produced at age x.

Note

The life table is calculated recursively until the age class (x) reaches xmax or survivorship (lx) falls below lx_crit — whichever comes first. To force calculation to xmax, set lx_crit = 0. Conversely, to force calculation to lx_crit, set xmax = Inf.

Starting from multiple stages

Rather than specifying argument start as a single stage class from which all individuals start life, it may sometimes be desirable to allow for multiple starting stage classes. For example, if the user wants to start the calculation of age-specific traits from reproductive maturity (i.e. first reproduction), the user should account for the possibility that there may be multiple stage classes in which an individual could first reproduce.

To specify multiple starting stage classes, specify argument start as the desired starting population vector (n1), giving the proportion of individuals starting in each stage class (the length of start should match the number of columns in the relevant MPM).

See function mature_distrib for calculating the proportion of individuals achieving reproductive maturity in each stage class.

References

Caswell, H. 2001. Matrix Population Models: Construction, Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

Caswell, H. 2006. Applications of Markov chains in demography. pp. 319-334 in A.N. Langville and W.J. Stewart (editors) MAM2006: Markov Anniversary Meeting. Boson Books, Raleigh, North Caroline, USA

Horvitz, C. & Tuljapurkar, S. 2008. Stage dynamics, period survival, and mortality plateaus. The American Naturalist 172: 203-2015. <doi:10.1086/589453>

Jones, O. R., Scheuerlein, A., Salguero-Gomez, R., Camarda, C. G., Schaible, R., Casper, B. B., Dahlgren, J. P., Ehrlén, J., García, M. B., Menges, E., Quintana-Ascencio, P. F., Caswell, H., Baudisch, A. & Vaupel, J. 2014. Diversity of ageing across the tree of life. Nature 505, 169-173. <doi:10.1038/nature12789>

See also

Author

Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>

Examples

data(mpm1)

mpm_to_table(matU = mpm1$matU, start = 2, xmax = 15)

#>     x          lx          dx        qx          Lx          Tx       ex
#> 1   0 1.000000000 0.500000000 0.5000000 0.750000000 2.000382820 2.000383
#> 2   1 0.500000000 0.183200000 0.3664000 0.408400000 1.250382820 2.500766
#> 3   2 0.316800000 0.108187000 0.3414994 0.262706500 0.841982820 2.657774
#> 4   3 0.208613000 0.064699290 0.3101403 0.176263355 0.579276320 2.776799
#> 5   4 0.143913710 0.043329919 0.3010826 0.122248751 0.403012965 2.800379
#> 6   5 0.100583791 0.029887091 0.2971363 0.085640246 0.280764215 2.791347
#> 7   6 0.070696700 0.020921617 0.2959348 0.060235892 0.195123969 2.760015
#> 8   7 0.049775083 0.014708992 0.2955091 0.042420587 0.134888077 2.709952
#> 9   8 0.035066091 0.010357661 0.2953754 0.029887261 0.092467490 2.636949
#> 10  9 0.024708430 0.007297162 0.2953309 0.021059849 0.062580229 2.532748
#> 11 10 0.017411268 0.005141838 0.2953167 0.014840349 0.041520380 2.384684
#> 12 11 0.012269430 0.003623311 0.2953121 0.010457775 0.026680031 2.174513
#> 13 12 0.008646119 0.002553291 0.2953106 0.007369474 0.016222256 1.876247
#> 14 13 0.006092829 0.001799274 0.2953101 0.005193192 0.008852782 1.452984
#> 15 14 0.004293555 0.001267929 0.2953100 0.003659590 0.003659590 0.852345
#> 16 15 0.003025625          NA        NA          NA          NA       NA

mpm_to_table(matU = mpm1$matU, start = "small", xmax = 15)  # equivalent using named life stages

#>     x          lx          dx        qx          Lx          Tx       ex
#> 1   0 1.000000000 0.500000000 0.5000000 0.750000000 2.000382820 2.000383
#> 2   1 0.500000000 0.183200000 0.3664000 0.408400000 1.250382820 2.500766
#> 3   2 0.316800000 0.108187000 0.3414994 0.262706500 0.841982820 2.657774
#> 4   3 0.208613000 0.064699290 0.3101403 0.176263355 0.579276320 2.776799
#> 5   4 0.143913710 0.043329919 0.3010826 0.122248751 0.403012965 2.800379
#> 6   5 0.100583791 0.029887091 0.2971363 0.085640246 0.280764215 2.791347
#> 7   6 0.070696700 0.020921617 0.2959348 0.060235892 0.195123969 2.760015
#> 8   7 0.049775083 0.014708992 0.2955091 0.042420587 0.134888077 2.709952
#> 9   8 0.035066091 0.010357661 0.2953754 0.029887261 0.092467490 2.636949
#> 10  9 0.024708430 0.007297162 0.2953309 0.021059849 0.062580229 2.532748
#> 11 10 0.017411268 0.005141838 0.2953167 0.014840349 0.041520380 2.384684
#> 12 11 0.012269430 0.003623311 0.2953121 0.010457775 0.026680031 2.174513
#> 13 12 0.008646119 0.002553291 0.2953106 0.007369474 0.016222256 1.876247
#> 14 13 0.006092829 0.001799274 0.2953101 0.005193192 0.008852782 1.452984
#> 15 14 0.004293555 0.001267929 0.2953100 0.003659590 0.003659590 0.852345
#> 16 15 0.003025625          NA        NA          NA          NA       NA

mpm_to_table(matU = mpm1$matU, matF = mpm1$matF, start = 2, xmax = 15)

#>     x          lx          dx        qx          Lx          Tx       ex
#> 1   0 1.000000000 0.500000000 0.5000000 0.750000000 2.000382820 2.000383
#> 2   1 0.500000000 0.183200000 0.3664000 0.408400000 1.250382820 2.500766
#> 3   2 0.316800000 0.108187000 0.3414994 0.262706500 0.841982820 2.657774
#> 4   3 0.208613000 0.064699290 0.3101403 0.176263355 0.579276320 2.776799
#> 5   4 0.143913710 0.043329919 0.3010826 0.122248751 0.403012965 2.800379
#> 6   5 0.100583791 0.029887091 0.2971363 0.085640246 0.280764215 2.791347
#> 7   6 0.070696700 0.020921617 0.2959348 0.060235892 0.195123969 2.760015
#> 8   7 0.049775083 0.014708992 0.2955091 0.042420587 0.134888077 2.709952
#> 9   8 0.035066091 0.010357661 0.2953754 0.029887261 0.092467490 2.636949
#> 10  9 0.024708430 0.007297162 0.2953309 0.021059849 0.062580229 2.532748
#> 11 10 0.017411268 0.005141838 0.2953167 0.014840349 0.041520380 2.384684
#> 12 11 0.012269430 0.003623311 0.2953121 0.010457775 0.026680031 2.174513
#> 13 12 0.008646119 0.002553291 0.2953106 0.007369474 0.016222256 1.876247
#> 14 13 0.006092829 0.001799274 0.2953101 0.005193192 0.008852782 1.452984
#> 15 14 0.004293555 0.001267929 0.2953100 0.003659590 0.003659590 0.852345
#> 16 15 0.003025625          NA        NA          NA          NA       NA
#>          mx       lxmx
#> 1   0.00000 0.00000000
#> 2  15.26600 7.63300000
#> 3  22.00578 6.97143000
#> 4  25.84823 5.39227780
#> 5  27.13917 3.90569858
#> 6  27.62496 2.77862313
#> 7  27.78025 1.96397169
#> 8  27.83270 1.38537517
#> 9  27.84948 0.97657248
#> 10 27.85497 0.68825265
#> 11 27.85673 0.48502105
#> 12 27.85730 0.34179324
#> 13 27.85749 0.24085915
#> 14 27.85755 0.16973126
#> 15 27.85756 0.11960798
#> 16 27.85757 0.08428657

### starting from first reproduction
repStages <- repro_stages(mpm1$matF)
n1 <- mature_distrib(matU = mpm1$matU, start = 2, repro_stages = repStages)
mpm_to_table(matU = mpm1$matU, start = n1)

#>     x           lx           dx        qx           Lx           Tx        ex
#> 1   0 1.000000e+00 3.242105e-01 0.3242105 0.8378947368 2.6787186124 2.6787186
#> 2   1 6.757895e-01 2.268500e-01 0.3356815 0.5623644737 1.8408238756 2.7239606
#> 3   2 4.489395e-01 1.360970e-01 0.3031522 0.3808909868 1.2784594019 2.8477322
#> 4   3 3.128425e-01 9.359475e-02 0.2991753 0.2660451262 0.8975684151 2.8690744
#> 5   4 2.192478e-01 6.496711e-02 0.2963182 0.1867641991 0.6315232889 2.8804094
#> 6   5 1.542806e-01 4.561886e-02 0.2956875 0.1314712165 0.4447590898 2.8827925
#> 7   6 1.086618e-01 3.210105e-02 0.2954217 0.0926112637 0.3132878733 2.8831467
#> 8   7 7.656074e-02 2.261206e-02 0.2953479 0.0652547114 0.2206766096 2.8823730
#> 9   8 5.394868e-02 1.593222e-02 0.2953217 0.0459825735 0.1554218982 2.8809211
#> 10  9 3.801646e-02 1.122679e-02 0.2953138 0.0324030715 0.1094393247 2.8787350
#> 11 10 2.678968e-02 7.911290e-03 0.2953111 0.0228340338 0.0770362532 2.8755945
#> 12 11 1.887839e-02 5.574982e-03 0.2953103 0.0160908977 0.0542022194 2.8711253
#> 13 12 1.330341e-02 3.928629e-03 0.2953100 0.0113390920 0.0381113217 2.8647791
#> 14 13 9.374777e-03 2.768465e-03 0.2953099 0.0079905451 0.0267722297 2.8557723
#> 15 14 6.606313e-03 1.950909e-03 0.2953099 0.0056308579 0.0187816846 2.8429906
#> 16 15 4.655403e-03 1.374787e-03 0.2953099 0.0039680099 0.0131508266 2.8248523
#> 17 16 3.280617e-03 9.687985e-04 0.2953099 0.0027962174 0.0091828167 2.7991130
#> 18 17 2.311818e-03 6.827027e-04 0.2953099 0.0019704668 0.0063865993 2.7625873
#> 19 18 1.629115e-03 4.810939e-04 0.2953099 0.0013885685 0.0044161326 2.7107549
#> 20 19 1.148022e-03 3.390221e-04 0.2953099 0.0009785105 0.0030275641 2.6372015
#> 21 20 8.089994e-04 2.389055e-04 0.2953099 0.0006895467 0.0020490536 2.5328246
#> 22 21 5.700939e-04 1.683544e-04 0.2953099 0.0004859167 0.0013595070 2.3847071
#> 23 22 4.017395e-04 1.186377e-04 0.2953099 0.0003424207 0.0008735903 2.1745190
#> 24 23 2.831019e-04 8.360278e-05 0.2953099 0.0002413005 0.0005311695 1.8762487
#> 25 24 1.994991e-04 5.891406e-05 0.2953099 0.0001700421 0.0002898690 1.4529842
#> 26 25 1.405850e-04 4.151615e-05 0.2953099 0.0001198270 0.0001198270 0.8523451
#> 27 26 9.906889e-05           NA        NA           NA           NA        NA