Calculate stochastic elasticities from a time-series of matrix population models and corresponding population vectors

Calculate stochastic elasticities given a time-series of matrix population models and corresponding population vectors, using the method described in Haridas et al. (2009).

perturb_stochastic(X_t, u_t)

Arguments

X_t

A list of matrix population models

u_t

A list of corresponding population vectors

Value

A list of three matrices:

E

matrix of stochastic elasticities

E_mu

matrix of stochastic elasticities to mean transition rates

E_sigma

matrix of stochastic elasticities to the variance in transition rates

References

Haridas, C. V., Tuljapurkar, S., & Coulson, T. 2009. Estimating stochastic elasticities directly from longitudinal data. Ecology Letters, 12, 806-812. <doi:10.1111/j.1461-0248.2009.01330.x>

See also

Author

Patrick Barks <patrick.barks@gmail.com>

Examples

# generate list of random MPMs
N <- 20 # number of years
s <- 3 # matrix dimension
X <- list() # matrix population model at time t
u <- list() # population vector at time t

for (t in 1:N) {
  X[[t]] <- matrix(runif(s^2), nrow = s, ncol = s)
}

# derive corresponding series of population vectors
u <- pop_vectors(X)

# calculate stochastic elasticities
perturb_stochastic(X, u)
#> $E
#>            [,1]      [,2]       [,3]
#> [1,] 0.09203099 0.1378016 0.09422870
#> [2,] 0.14811891 0.1693348 0.09288703
#> [3,] 0.09040402 0.1033997 0.07179426
#> 
#> $E_mu
#>            [,1]      [,2]       [,3]
#> [1,] 0.09622213 0.1423163 0.11586029
#> [2,] 0.14771839 0.1720908 0.09948410
#> [3,] 0.10020166 0.1155084 0.08073676
#> 
#> $E_sigma
#>               [,1]         [,2]         [,3]
#> [1,] -0.0041911351 -0.004514756 -0.021631600
#> [2,]  0.0004005197 -0.002756072 -0.006597069
#> [3,] -0.0097976348 -0.012108709 -0.008942501
#>